% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundPoisson.R
\name{compoundPoisson}
\alias{compoundPoisson}
\title{Function for estimating a various compound Poisson regression models where the error term follows specific distributions using Maximum Simulated Likelihood when there is not a closed form PMF}
\usage{
compoundPoisson(
  formula,
  data,
  distrib = "GE",
  ndraws = 1500,
  method = "BHHH",
  max.iters = 1000
)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}.}

\item{distrib}{a distribution used with the Poisson distribution to create the compound Poisson distribution. Options include:
\code{"GE"} for the Generalized Exponential distribution (resulting in the Poisson-Generalized-Exponential model),
\code{"W"} for the weibull distribution (resulting in the Poisson-Weibull model),
\code{"IG"} for the Inverse Gamma distribution (resulting in the Poisson-Inverse-Gamma model)
\code{"Weibull"} for the weibull distribution (resulting in the Poisson-Weibull model)}

\item{ndraws}{the number of Halton draws to use for estimating the random parameters,}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a various compound Poisson regression models where the error term follows specific distributions using Maximum Simulated Likelihood when there is not a closed form PMF
}
\note{
For the distributions used for compounding with the Poisson, the CDF is used to determine the quantile function. These include:

Generalized Exponential Distribution CDF and Quantile Functions (see Gupta & Kundu (2007))
\deqn{F(x)=\frac{\left(e^{\beta x}-1\right)^\alpha-1}{\alpha\beta}}
\deqn{Q(p)=\frac{ln\left(\left(P\cdot\alpha+1\right)^{\frac{1}{\alpha}}+1\right)}{\beta}}
Where \eqn{\beta} is the shape parameter, \eqn{\alpha} is the scale parameter, and \eqn{\P} is the percentile. The parameter constraints are \eqn{\tau>0} and \eqn{\alpha>0}.
    The mean and variance of this distribution are:
\deqn{\mu=\frac{1}{\beta}\left(\psi(\alpha+1)-\psi(1)\right)}
\deqn{\sigma^2=\frac{1}{\beta}\left(\psi\prime(1)-\psi\prime(\alpha+1)\right)}
Where \eqn{\psi} is the digamma function and \eqn{\psi\prime} is the trigamma function
    The resulting Poisson-Generalized Exponential distribution has the mean and variance:
\deqn{E[Y]=\mu=e^{X\beta}}
\deqn{Var[y]=\sigma^2=\mu+\left(\frac{\psi\prime(1)-\psi\prime(\alpha+1)}{\beta\left(\psi(\alpha+1)-\psi(1)\right)^2}\right)\mu^2}

Weibull Distribution  CDF and Quantile Functions
\deqn{F(x)=1-exp(-(x/\beta)^\alpha)}
\deqn{Q(p)=\beta(-ln(1-P))^{1/\alpha}}
Where \eqn{\beta} is the shape parameter, \eqn{\alpha} is the scale parameter, and \eqn{P} is the percentile. There are constraints for \eqn{\alpha>0} and \eqn{\beta>0}.
    The mean and variance of this distribution are:
\deqn{\mu=\beta\Gamma(1+1/\alpha)}
\deqn{\sigma^2=\beta^2\left(\Gamma(1+2/\alpha)-\left(\Gamma(1+1/\alpha)\right)^2\right)}
    The resulting Poisson-Weibull distribution has the mean and variance:
\deqn{E[Y]=\mu=e^{X\beta}}
\deqn{Var[y]=\sigma^2=\mu+\left(\frac{\Gamma\left(\frac{\alpha+2}{\alpha}\right)}{\Gamma\left(1+\frac{1}{\alpha}\right)^2}-1\right)\mu^2}

Inverse-Gamma  CDF Function
\deqn{F(x)=\frac{\Gamma\left(\alpha,\frac{\beta}{x}\right)}{\Gamma(\alpha)}}
Where \eqn{\beta} is the shape parameter and \eqn{\alpha} is the scale parameter. For the variance to be defined, \eqn{\alpha>2}. Additionally, an additional constraint is \eqn{\beta>0}.
    The mean and variance of this distribution are:
\deqn{\mu=\frac{\beta}{\alpha-1}}
\deqn{\sigma^2=\frac{\beta^2}{(\alpha-1)^2(\alpha-2)}}
    The resulting Poisson-Inverse-Gamma distribution has the mean and variance:
\deqn{E[Y]=\mu=e^{X\beta}}
\deqn{Var[y]=\sigma^2=\mu+\frac{\beta^2}{\alpha-2}\mu^2}
}
\examples{
\donttest{

## Poisson-Generalized-Exponential Model
data("washington_roads")
poisgenexp.mod <- compoundPoisson(Total_crashes ~ lnaadt + lnlength,
                                data=washington_roads,
                                distrib="GE",
                                ndraws=500, method="bfgs")
summary(poisgenexp.mod)

## Poisson-Weibull Model
poisweib.mod <- compoundPoisson(Total_crashes ~ lnaadt + lnlength,
                                data=washington_roads,
                                distrib="W",
                                ndraws=500, method="bfgs")
summary(poisweib.mod)

## Poisson-Inverse-Gamma Model
poisginvgamma.mod <- compoundPoisson(Total_crashes ~ lnaadt + lnlength,
                                data=washington_roads,
                                distrib="IG",
                                ndraws=500, method="bfgs")
summary(poisginvgamma.mod)}
}
\references{
Gupta, R. D., & Kundu, D. (2007). Generalized exponential distribution: Existing results and some recent developments. Journal of Statistical planning and inference, 137(11), 3537-3547.
}
