% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Generate a covariance matrix using a correlation matrix and vector of standard deviations}
\usage{
cor2cov(C, S)
}
\arguments{
\item{C}{A correlation matrix.}

\item{S}{A vector of standard deviations.}
}
\value{
A covariance matrix
}
\description{
Generate a covariance matrix using a correlation matrix and vector of standard deviations
}
\examples{
C <- matrix(c(1,-0.3,0.7,-0.3,1,-0.2,0.7,-0.2,1), 3, 3)
S <- c(0.5, 2, 1.25)
cor2cov(C,S)
}
