% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexCountRegClass.R
\docType{class}
\name{flexCountReg-class}
\alias{flexCountReg-class}
\title{flexCountReg Class}
\description{
A class to represent various objects created by the flexCountReg package.
}
\section{Slots}{

\describe{
\item{\code{model}}{The fitted model object (can be any type).}

\item{\code{data}}{The data used for fitting the model (data frame).}

\item{\code{formula}}{The R formula used in the regression model (main formula, not
including random parameters, etc.)}

\item{\code{call}}{The matched call (language).}

\item{\code{additional}}{Any additional information (list).}
}}

