% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genWaring.R
\name{genWaring}
\alias{genWaring}
\title{Function for estimating a Generalized Waring regression model}
\usage{
genWaring(formula, data, method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in  `formula`.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}} documentation.}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a Generalized Waring regression model
}
\details{
# Generalized Waring Probability Mass Function, Mean, and Variance
The following are the versions of the PMF, mean, and variance used in this function. This is adjusted from the typical formulation by replacing parameter \code{k} with \eqn{\mu}
\deqn{PMF=\frac{\Gamma(\alpha+y)\Gamma(k+y)\Gamma(\rho+k)\Gamma(\alpha+\rho)}{y!\Gamma(\alpha)\Gamma(k)\Gamma(\rho)\Gamma(\alpha+k+\rho+y)}}
\deqn{\mu=e^{X\beta}=\frac{\alpha k}{\rho-1}}
\deqn{\sigma^2=\frac{\alpha k(\alpha+k+\rho-1)}{(\rho-1)^2(\rho-2)}}

The distribution parameters are often considered to capture the randomness (parameter \deqn{\alpha}), proneness (parameter \deqn{}k), and liability (parameter \deqn{\rho}) of the data.

#' If we use:
\deqn{\alpha=\frac{\mu k}{\rho-1}}

The PMF becomes:

\deqn{PMF=\frac{\Gamma\left(\frac{\mu k}{\rho-1}+y\right)\Gamma(k+y)\Gamma(\rho+k)\Gamma\left(\frac{\mu k}{\rho-1}+\rho\right)}{y!\Gamma\left(\frac{\mu k}{\rho-1}\right)\Gamma(k)\Gamma(\rho)\Gamma\left(\frac{\mu k}{\rho-1}+k+\rho+y\right)}}

#' This results in a regression model where:
\deqn{\mu=e^{X\beta}}
\deqn{\sigma^2=\frac{\mu k^2\left(\frac{\mu k}{\rho-1}+k+\rho-1\right)}{(\rho-1)^3(\rho-2)}=\left(\frac{k^3+\rho k^2- k^2}{(\rho-1)^3(\rho-2)}\right)\mu+\left(\frac{k^3}{(\rho-1)^4(\rho-2)}\right)\mu^2}

Note that when \deqn{p=1} or \deqn{p=2}, the distribution is undefined.
}
\examples{
\donttest{

# Generalized Waring Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
genwaring.mod <- genWaring(Total_crashes ~ lnaadt + lnlength + speed50 +
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                method='BFGS')
summary(genwaring.mod)}
}
