% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chol.R
\name{get_chol}
\alias{get_chol}
\title{Create Cholesky Matrix from Parameters in a Random Parameter Model}
\usage{
get_chol(pars, Nvars)
}
\arguments{
\item{pars}{A numeric vector of parameters to include in the Cholesky matrix.}

\item{Nvars}{An integer specifying the number of random variables for the Cholesky matrix.}
}
\value{
A matrix with \code{Nvars} columns and rows containing the upper Cholesky matrix.
}
\description{
This function creates a Cholesky matrix from parameters in a random parameter model.
}
