% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.predict.R
\name{mle.predict}
\alias{mle.predict}
\title{Predictions for non-random parameters count models}
\usage{
mle.predict(model, data)
}
\arguments{
\item{model}{a model object estimated using any of the models other than the random parameters (\code{rpnb}).}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe.}
}
\description{
Predictions for non-random parameters count models
}
\examples{
\donttest{

## Standard Negative Binomial
data("washington_roads")
nbg <- nbg(Total_crashes ~ - 1 + lnlength + speed50,
           rpar_formula = ~ lnaadt,
           data = washington_roads)
summary(nbg)

predictions <- mle.predict(nbg, washington_roads)
print(predictions)}
}
