% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbg.R
\name{nbg}
\alias{nbg}
\title{Function for estimating a variety of negative binomial models (NB-1, NB-2,
NB-P and generalized versions of each)}
\usage{
nbg(
  formula,
  data,
  form = "nb2",
  ln.alpha.formula = NULL,
  method = "BHHH",
  max.iters = 200
)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula}
and \code{rpar_formula}. This can be the data used for estimating the model
or another dataframe,}

\item{form}{the version of the negative binomial to estimate (\code{"nb2"}
estimates the NB-2, \code{"nb1"} estimates the NB-1, \code{"nbp"} estimates
the NB-P)}

\item{ln.alpha.formula}{an optional formula for using independent variables
to estimate the natural log of the overdispersion parameter (makes the
model a generalized negative binomial).}

\item{method}{a method to use for optimization in the maximum likelihood
estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization
method to perform.}
}
\description{
Function for estimating a variety of negative binomial models (NB-1, NB-2,
NB-P and generalized versions of each)
}
\details{
The NB-1, NB-2, and NB-P versions of the negative binomial distribution are 
based on Greene (2008).  The details of each of these are provided below.

# NB-1 Distribution
The PMF and log-likelihood functions are:
\deqn{P(Y = y) = \frac{\Gamma(y + \frac{\mu}{\alpha})}{y! \, \Gamma(\frac{\mu}{\alpha})} \left( \frac{\frac{\mu}{\alpha}}{\frac{\mu}{\alpha} + \mu} \right)^{\frac{\mu}{\alpha}} \left( \frac{\mu}{\frac{\mu}{\alpha} + \mu} \right)^y}
\deqn{LL_{\text{NB1}}(\beta, \alpha) = \sum_{i=1}^n \left[ \ln \Gamma\left( y_i + \frac{\mu_i}{\alpha} \right) - \ln \Gamma\left( \frac{\mu_i}{\alpha} \right) - \ln y_i! + \frac{\mu_i}{\alpha} \ln \left( \frac{\frac{\mu_i}{\alpha}}{\frac{\mu_i}{\alpha} + \mu_i} \right) + y_i \ln \left( \frac{\mu_i}{\frac{\mu_i}{\alpha} + \mu_i} \right) \right]}

The Gradients are:
\deqn{\frac{\partial LL_{\text{NB1}}}{\partial \beta} = \sum_{i=1}^n \left[ \left( \psi\left(y_i + \frac{\mu_i}{\alpha}\right) - \psi\left(\frac{\mu_i}{\alpha}\right) + \ln \left( \frac{\frac{\mu_i}{\alpha}}{\frac{\mu_i}{\alpha} + \mu_i} \right) - \frac{y_i}{\frac{\mu_i}{\alpha} + \mu_i} \right) \cdot \mu_i \cdot X_i \right]}
\deqn{\frac{\partial LL_{\text{NB1}}}{\partial \eta} = \alpha \sum_{i=1}^n \left[ \psi\left(y_i + \frac{\mu_i}{\alpha}\right) - \psi\left(\frac{\mu_i}{\alpha}\right) + \ln \left( \frac{\frac{\mu_i}{\alpha}}{\frac{\mu_i}{\alpha} + \mu_i} \right) - \frac{\mu_i}{\frac{\mu_i}{\alpha} + \mu_i} \right]}

# NB-2 Distribution
The PMF and log-likelihood functions are:
\deqn{P(Y = y) = \frac{\Gamma(y + \alpha)}{y! \, \Gamma(\alpha)} \left( \frac{\alpha}{\alpha + \mu} \right)^\alpha \left( \frac{\mu}{\alpha + \mu} \right)^y}
\deqn{LL_{\text{NB2}} = \sum_{i=1}^n \left[ \ln \Gamma(y_i + \alpha) - \ln \Gamma(\alpha) - \ln y_i! + \alpha \ln \left( \frac{\alpha}{\alpha + \mu_i} \right) + y_i \ln \left( \frac{\mu_i}{\alpha + \mu_i} \right) \right]}

The Gradients are:
\deqn{\frac{\partial LL_{\text{NB2}}}{\partial \beta} = \sum_{i=1}^n \left[ \left( \psi(y_i + \alpha) - \psi(\alpha) + \ln \left( \frac{\alpha}{\alpha + \mu_i} \right) - \frac{y_i}{\alpha + \mu_i} \right) \cdot \mu_i \cdot X_i \right]}
\deqn{\frac{\partial LL_{\text{NB2}}}{\partial \eta} = \alpha \sum_{i=1}^n \left[ \psi(y_i + \alpha) - \psi(\alpha) + \ln \left( \frac{\alpha}{\alpha + \mu_i} \right) - \frac{\mu_i}{\alpha + \mu_i} \right]}

# NB-P distribution
The PMF and log-likelihood functions are:
\deqn{P(Y = y) = \frac{\Gamma(y + \frac{\mu^{2-p}}{\alpha})}{y! \, \Gamma(\frac{\mu^{2-p}}{\alpha})} \left( \frac{\frac{\mu^{2-p}}{\alpha}}{\frac{\mu^{2-p}}{\alpha} + \mu} \right)^{\frac{\mu^{2-p}}{\alpha}} \left( \frac{\mu}{\frac{\mu^{2-p}}{\alpha} + \mu} \right)^y}
\deqn{LL_{\text{NBP}}(\beta, \alpha, p) = \sum_{i=1}^n \left[ \ln \Gamma\left( y_i + \frac{\mu_i^{2-p}}{\alpha} \right) - \ln \Gamma\left( \frac{\mu_i^{2-p}}{\alpha} \right) - \ln y_i! + \frac{\mu_i^{2-p}}{\alpha} \ln \left( \frac{\frac{\mu_i^{2-p}}{\alpha}}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) + y_i \ln \left( \frac{\mu_i}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) \right]}

The Gradients are:
\deqn{\frac{\partial LL_{\text{NBP}}}{\partial \beta} = \sum_{i=1}^n \left[ \left( \psi\left(y_i + \frac{\mu_i^{2-p}}{\alpha}\right) - \psi\left(\frac{\mu_i^{2-p}}{\alpha}\right) + \ln \left( \frac{\frac{\mu_i^{2-p}}{\alpha}}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) - \frac{y_i}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) \cdot \mu_i \cdot X_i \right]}
\deqn{\frac{\partial LL_{\text{NBP}}}{\partial \eta} = \alpha \sum_{i=1}^n \left[ \psi\left(y_i + \frac{\mu_i^{2-p}}{\alpha}\right) - \psi\left(\frac{\mu_i^{2-p}}{\alpha}\right) + \ln \left( \frac{\frac{\mu_i^{2-p}}{\alpha}}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) - \frac{\mu_i^{2-p}}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right]}
\deqn{\frac{\partial LL_{\text{NBP}}}{\partial p} = \sum_{i=1}^n \left[ \left( \psi\left(y_i + \frac{\mu_i^{2-p}}{\alpha}\right) - \psi\left(\frac{\mu_i^{2-p}}{\alpha}\right) + \ln \left( \frac{\frac{\mu_i^{2-p}}{\alpha}}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) - \frac{\mu_i}{\frac{\mu_i^{2-p}}{\alpha} + \mu_i} \right) \cdot \frac{\partial}{\partial p} \left( \frac{\mu_i^{2-p}}{\alpha} \right) \right]}
}
\examples{

## NB-P model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0)

nbp.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                    ShouldWidth04 + AADTover10k,
                    data=washington_roads, form = 'nbp', method = 'NM',
                    max.iters=3000)
summary(nbp.base)

## Generalized NB-P model

nbp.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                form = 'nbp',
                                method = 'NM',
                                max.iters=3000,
                                ln.alpha.formula = ~ 1+lnlength)
summary(nbp.overdispersion)

## NB-1 Model
nb1.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb1',
                        method = 'NM',
                        max.iters=3000)
summary(nb1.base)

## Generalize NB-1 Model
nb1.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb1',
                        method = 'NM',
                        max.iters=3000, ln.alpha.formula = ~ 1+lnlength)
summary(nb1.overdispersion)

## NB-2 Model
nb2.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb2',
                        method = 'NM',
                        max.iters=3000)
summary(nb2.base)

## Generalize NB-2 Model
nb2.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb2',
                        method = 'NM',
                        max.iters=3000, ln.alpha.formula = ~ 1+lnlength)
summary(nb2.overdispersion)

}
\references{
Greene, W. (2008). Functional forms for the negative binomial model for count data. Economics Letters, 99(3), 585-590.
}
