% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGE.R
\name{poisGE}
\alias{poisGE}
\title{Poisson-Generalized-Exponential Regression}
\usage{
poisGE(
  formula,
  data,
  ln.scale.formula = NULL,
  ndraws = 1500,
  method = "BHHH",
  max.iters = 200,
  print.level = 0
)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe,}

\item{ln.scale.formula}{an optional formula for using independent variables to estimate the natural log of the scale parameter.}

\item{ndraws}{the number of Halton draws to use for the integration over the lognormal distribution.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}

\item{print.level}{determines the level of verbosity for printing details of the optimization as it is computed. A value of 0 does not print out any information, a value of 1 prints minimal information, and a value of 2 prints the most information.}
}
\description{
Poisson-Generalized-Exponential Regression
}
\details{
This implements maximum simulated likelihood (MSL) to estimate a Poisson-Generalized Exponential regression model. The regression model has the flexibility to model the scale parameter as a function of independent variables, similar to the generalized negative binomial.

Details of the distribution can be found with the function \code{\link{dpge}}
The t-statistics and p-values for the coefficients related to ln(sigma) are,
by default, testing if the coefficients are different from a value of 0. This
has little practical meaning given that they are coefficients for ln(sigma).
They are not testing if the coefficients have statistical significance in
terms of improvement over a Poisson model. The Likelihood-Ratio test results
provided in the output provide a test comparing if the Poisson-Lognormal
model provides a statistically significant improvement in model fit over the
Poisson model.
}
\examples{
\donttest{

## Generalized Poisson-Lognormal
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0)

poisge.mod <- poisGE(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        ln.scale.formula = ~ speed50,
                        data=washington_roads, 
                        ndraws = 100, 
                        method = 'nm')
summary(poisge.mod)
}
}
