% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisInvGaus.R
\name{poisInvGaus}
\alias{poisInvGaus}
\title{Function for estimating a Poisson-Inverse-Gaussian regression model}
\usage{
poisInvGaus(formula, data, form = "Type 1", method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula}.}

\item{form}{optional parameter indicating which formulation to use. Options include "Type 1" which is the standard form (and is the default) or "Type 2" which follows the formulation by Dean et. al. (1987).}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a Poisson-Inverse-Gaussian regression model
}
\details{
The Poisson-Inverse-Gaussian regression model is based on the Poisson-Inverse-Gaussian Distribution.
}
\examples{
\donttest{

## Poisson-Inverse-Gaussian Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable

poisinvgaus.mod <- poisInvGaus(Total_crashes ~ lnaadt + lnlength + speed50 + 
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                method="nm",
                                max.iters = 1000)
summary(poisinvgaus.mod)}
}
\seealso{
[dpinvgaus()] for additional details of the distribution.

The expected value of the distribution in the regression utilizes a log-link function. Thus, the mean is:
\deqn{\mu=e^{X\beta}}

The variance function for the Type 1 distribution (which is the default) is:
\deqn{\sigma^2=\mu+\eta\mu}

While the variance for the Type 2 distribution is:
\deqn{\sigma^2=\mu+\eta\mu^2}

The parameter \eqn{\eta} is estimated as the natural logarithm transformed value, ln(eta), to ensure that \eqn{\eta>0}.
}
