% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisLind.R
\name{poisLind}
\alias{poisLind}
\title{Function for estimating a Poisson-Lindley regression model}
\usage{
poisLind(formula, data, method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula}.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform,}
}
\description{
Function for estimating a Poisson-Lindley regression model
}
\details{
The Poisson-Lindley regression is based on a compound Poisson-Lindley distribution. It handles count outcomes with high levels of zero observations (or other high densities at low outcome values) that standard count regression methods, including the negative binomial, may struggle to adequately capture or model.

The compound Probability Mass Function(PMF) for the Poisson-Lindley (PL) distribution is:
\deqn{f(y|\theta,\lambda)=\frac{\theta^2\lambda^y(\theta+\lambda+ y+1)}{(\theta+1)(\theta+\lambda)^{y+2}}}

Where \eqn{\theta} and \eqn{\lambda} are distribution parameters with the restrictions that \eqn{\theta>0} and \eqn{\lambda>0}, and \eqn{y} is a non-negative integer.

The expected value of the distribution is:
\deqn{\mu=\frac{\lambda(\theta+2)}{\theta(\theta+1)}}

If a log-link function is used, the mean is:
\deqn{\mu=e^{X\beta}=\frac{\lambda(\theta+2)}{\theta(\theta+1)}}

Thus, the parameter \eqn{\lambda} in the PL distribution when applied to regression analysis is:
\deqn{\lambda=\frac{\mu\theta(\theta+1)}{\theta+2}}

The variance function is defined as:#'
\deqn{\sigma^2=\mu+\left(1-\frac{2}{(\theta+2)^2}\right)\mu^2}

It should be noted that the p-value for the parameter `ln(theta)` in the model summary is testing if the parameter `theta` is equal to a value of 1. This has no practical meaning. The Likelihood-Ratio (LR) test compares the Poisson-Lindley regression with a Poisson regression with the same independent variables. Thus, the PR test result indicates the statistical significance for the improvement in how well the model fits the data over a Poisson regression. This indicates the statistical significance of the `theta` parameter.
}
\examples{
\donttest{

## Poisson-Lindley Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislind.mod <- poisLind(Animal ~ lnaadt + lnlength + speed50 + ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                method="nm",
                                max.iters = 1000)
summary(poislind.mod)}
}
