% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisLindGamma.R
\name{poisLindGamma}
\alias{poisLindGamma}
\title{Function for estimating a Poisson-Lindley-Gamma (i.e., Negative Binomial-Lindley) regression model}
\usage{
poisLindGamma(formula, data, method = "BHHH", ndraws = 1500, max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}}.}

\item{ndraws}{the number of Halton draws to use for the integration over the gamma distribution.}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a Poisson-Lindley-Gamma (i.e., Negative Binomial-Lindley) regression model
}
\details{
The Poisson-Lindley-Gamma regression is based on a compound Poisson-Lindley-Gamma distribution. Details of the distribution can be seen at \code{\link[flexCountReg]{dplindGamma}}.

The mean for the regression model is:
\deqn{\mu=e^{X\beta}}

The variance function is defined as:
\deqn{\sigma^2=\mu+\left(\alpha+1-\frac{2}{(\theta+2)^2}\right)\mu^2}

It should be noted that the p-value for the parameters `ln(theta)` and `ln(alpha)` in the model summary are testing if the parameter `theta` and `alpha` are equal to a value of 1.
}
\examples{
\donttest{

## Poisson-Lindley-Gamma Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislindgamma.mod <- poisLindGamma(Animal ~ lnaadt + lnlength + speed50 +
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                method="nm",
                                ndraws=100,
                                max.iters = 1000)
summary(poislindgamma.mod)}
}
