% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisLindLnorm.R
\name{poisLindLnorm}
\alias{poisLindLnorm}
\title{Function for estimating a Poisson-Lindley-Lognormal regression model}
\usage{
poisLindLnorm(formula, data, method = "BHHH", ndraws = 1500, max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}}.}

\item{ndraws}{the number of Halton draws to use for the integration over the gamma distribution.}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a Poisson-Lindley-Lognormal regression model
}
\details{
The Poisson-Lindley-Lognormal regression is based on a compound Poisson-Lindley-Lognormal distribution. Details of the distribution can be seen at \code{\link[flexCountReg]{dplindLnorm}}.

The mean for the regression model is:
\deqn{\mu=e^{X\beta}}

The variance function is defined as:
\deqn{\sigma^2=\mu+\left(\frac{1-\frac{2}{(\theta+2)^2}}{e^{\frac{\sigma^2}{2}}}+e^{\sigma^2}-1\right)\mu^2}

It should be noted that the p-value for the parameters `ln(theta)` and `ln(sigma)` in the model summary are testing if the parameter `theta` and `sigma` are equal to a value of 1.
}
\examples{
\donttest{

## Poisson-Lindley-Lognormal Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislindlnorm.mod <- poisLindLnorm(Animal ~ lnaadt + lnlength + speed50 
                                        + ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                method="nm",
                                ndraws=100,
                                max.iters = 1000)
summary(poislindlnorm.mod)}
}
