% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisLogn.R
\name{poisLogn}
\alias{poisLogn}
\title{Function for estimating a Poisson Lognormal regression model via maximum simulated likelihood}
\usage{
poisLogn(formula, data, ndraws = 1500, method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}.}

\item{ndraws}{the number of Halton draws to use for estimating the random parameters,}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform,}
}
\description{
Function for estimating a Poisson Lognormal regression model via maximum simulated likelihood
}
\examples{
\donttest{

## Poisson-Lognormal Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislog.mod <- poisLogn(Total_crashes ~ lnaadt + lnlength + speed50 + ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                ndraws=50)
summary(poislog.mod)}
}
