% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plindLnorm.R
\name{Poisson-Lindley-Lognormal}
\alias{Poisson-Lindley-Lognormal}
\alias{dplindLnorm}
\alias{pplindLnorm}
\alias{qplindLnorm}
\alias{rplindLnorm}
\title{Poisson-Lindley-Lognormal Distribution}
\usage{
dplindLnorm(
  x,
  mean = 1,
  theta = 1,
  sigma = 1,
  lambda = NULL,
  ndraws = 1500,
  log = FALSE
)

pplindLnorm(
  q,
  mean = 1,
  theta = 1,
  lambda = NULL,
  sigma = 1,
  ndraws = 1500,
  lower.tail = TRUE,
  log.p = FALSE
)

qplindLnorm(p, mean = 1, theta = 1, sigma = 1, ndraws = 1500, lambda = NULL)

rplindLnorm(n, mean = 1, theta = 1, sigma = 1, ndraws = 1500, lambda = NULL)
}
\arguments{
\item{x}{numeric value or a vector of values.}

\item{mean}{numeric value or vector of mean values for the distribution (the values have to be greater than 0).}

\item{theta}{single value or vector of values for the theta parameter of the distribution (the values have to be greater than 0).}

\item{sigma}{single value or vector of values for the sigma parameter of the lognormal distribution (the values have to be greater than 0).}

\item{lambda}{alternative parameterization (use instead of the mean); numeric value or vector of values for lambda parameter of the distribution (the values have to be greater than 0).}

\item{ndraws}{the number of Halton draws to use for the integration.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{quantile or a vector of quantiles.}

\item{lower.tail}{logical; if TRUE, probabilities p are \eqn{P[X\leq x]} otherwise, \eqn{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{probability or a vector of probabilities.}

\item{n}{the number of random numbers to generate.}
}
\description{
These functions provide density, distribution function, quantile function, and random number generation for the Poisson-Lindley-Lognormal (PLL) Distribution
}
\details{
The Poisson-Lindley-Lognormal is a 3-parameter count distribution that captures high densities for small integer values and provides flexibility for heavy tails.


\code{dplind} computes the density (PDF) of the Poisson-Lindley Distribution.

\code{pplind} computes the CDF of the Poisson-Lindley Distribution.

\code{qplind} computes the quantile function of the Poisson-Lindley Distribution.

\code{rplind} generates random numbers from the Poisson-Lindley Distribution.

The compound Probability Mass Function(PMF) for the Poisson-Lindley-Lognormal (PLL) distribution is:
\deqn{f(y|\mu,\theta,\alpha)=\int_0^\infty \frac{\theta^2\mu^y x^y(\theta+\mu\cdot x+y+1)}{(\theta+1)(\theta+\mu\cdot x)^{y+2}} \frac{exp\left(-\frac{ln^2(x)}{2\sigma^2} \right)}{x\sigma\sqrt{2\pi}}dx}

Where \eqn{\theta} and \eqn{\lambda} are distribution parameters from the Poisson-Lindley distribution with the restrictions that \eqn{\theta>0} and \eqn{\lambda>0}, \eqn{\sigma} is a parameter for the lognormal distribution with the restriction \eqn{\sigma>0}, and \eqn{y} is a non-negative integer.

The expected value of the distribution is:
\deqn{E[y]=\mu=e^{X\beta}=\frac{\lambda(\theta+2)e^{\frac{\sigma^2}{2}}}{\theta(\theta+1)}}

The default is to use the input mean value for the distribution. However, the lambda parameter can be used as an alternative to the mean value.

Halton draws are used to perform simulation over the lognormal distribution to solve the integral.
}
\examples{
dplindLnorm(0, mean=0.75, theta=7, sigma=2, ndraws=2000)
pplindLnorm(c(0,1,2,3,5,7,9,10), mean=0.75, theta=7, sigma=2, ndraws=500)
qplindLnorm(c(0.1,0.3,0.5,0.9,0.95), lambda=4.67, theta=7, sigma=2, ndraws=500)
rplindLnorm(30, mean=0.75, theta=7, sigma=2, ndraws=500)

}
