% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.nbg.R
\name{predict.nbg}
\alias{predict.nbg}
\title{Function for generating predictions based on the random parameters negative binomial with multiple optional methods}
\usage{
\method{predict}{nbg}(model, data)
}
\arguments{
\item{model}{a model object estimated using the \code{rpnb} function,}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe,}

\item{method}{the method to be used in generating the predictions (options include \code{Simulated}, \code{Approximate}, or \code{Individual}).}
}
\description{
Function for generating predictions based on the random parameters negative binomial with multiple optional methods
}
\note{
the method option \code{Individual} requires that the outcome be observed for all observations.
}
\examples{
\donttest{

## Random Parameters Negative Binomial model
data("washington_roads")
nb <- rpnb(Total_crashes ~ - 1 + lnlength + speed50,
           rpar_formula = ~ lnaadt,
           data = washington_roads)
summary(nb)

rpnb.predictions <- predict.rpnb(nb, washington_roads, method=Individual)
print(rpnb.predictions)}
}
