% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.poisLind.R
\name{predict.poisLind}
\alias{predict.poisLind}
\title{Function for generating predictions based on the Poisson Lindley model}
\usage{
\method{predict}{poisLind}(model, data)
}
\arguments{
\item{model}{a Poisson-Lindley model object,}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe,}

\item{method}{the method to be used in generating the predictions (options include \code{Simulated}, \code{Approximate}, or \code{Individual}).}
}
\description{
Function for generating predictions based on the Poisson Lindley model
}
\note{
the method option \code{Individual} requires that the outcome be observed for all observations.
}
\examples{
\donttest{

## Poisson-Lindley Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislind.mod <- poisLind(Total_crashes ~ lnaadt + lnlength + speed50 + ShouldWidth04 + AADTover10k,
                                data=washington_roads)
summary(poislind.mod)

predicted <- predict.poisLind(poislind.mod, washington_roads)
print(predicted)}
}
