% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.poisLogn.R
\name{predict.poisLogn}
\alias{predict.poisLogn}
\title{Function for generating predictions based on the Poisson Lognormal model}
\usage{
\method{predict}{poisLogn}(model, data)
}
\arguments{
\item{model}{a Poisson-Lognormal model object,}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe,}

\item{method}{the method to be used in generating the predictions (options include \code{Simulated}, \code{Approximate}, or \code{Individual}).}
}
\description{
Function for generating predictions based on the Poisson Lognormal model
}
\note{
the method option \code{Individual} requires that the outcome be observed for all observations.
}
\examples{
\donttest{

## Poisson-Lognormal Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
poislog.mod <- poisLogn(Total_crashes ~ lnaadt + lnlength + speed50 + ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                ndraws=50)
summary(poislog.mod)

predicted <- predict.poisLogn(poislog.mod, washington_roads)
print(predicted)}

}
