% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCompTest.R
\name{regCompTest}
\alias{regCompTest}
\title{Compare Regression Models with Likelihood Ratio Test, AIC, and BIC}
\usage{
regCompTest(
  model,
  data = NULL,
  basemodel = "Poisson",
  variables = FALSE,
  print = FALSE,
  ...,
  model2 = NULL
)
}
\arguments{
\item{model}{A fitted regression model object.}

\item{data}{An options data frame containing the variables in the model. If
not supplied, the original data used to estimate the model will be used.}

\item{basemodel}{A character string specifying the family of base model to
compare against (options include the family from \code{\link{countreg}} or
"Poisson"). Default is "Poisson".}

\item{variables}{Logical. If \code{TRUE}, the base model will include the
same variables as the provided model. If \code{FALSE}, the base model will
be an intercept-only model. Default is \code{FALSE}.}

\item{print}{Logical. If \code{TRUE}, a table of the results will be shown.
If \code{FALSE}, the table of results will not be printed to the console.}

\item{...}{Additional arguments to be passed to the base model fitting
function - options are any argument from the \code{\link{countreg}}
function.}

\item{model2}{An optional second fitted regression model object to compare.
If supplied, the function returns comparison results for both trained
models in addition to the base-model comparison.}
}
\value{
A list containing the following components.
\itemize{
\item For the original single-model workflow: \code{LL}, \code{LLbase},
\code{LR}, \code{LRdof}, \code{AIC}, \code{AICbase}, \code{BIC},
\code{BICbase}, \code{LR_pvalue}, \code{PseudoR2}, \code{statistics},
\code{gtTable}, \code{latexTable}, and \code{htmlTable}.
\item If \code{model2} is supplied: \code{model1}, \code{model2}, and a
combined \code{statistics} table plus the associated table objects.
}
}
\description{
This function compares fitted regression model objects using the
Likelihood Ratio (LR) test, Akaike Information Criterion (AIC), and Bayesian
Information Criterion (BIC).
}
\details{
The function performs the following steps:
\enumerate{
\item Fits the base model, either a Poisson regression or another specified
model.
\item Computes the log-likelihoods of the provided model(s) and the base
model.
\item Calculates the AIC and BIC for the model(s).
\item Conducts a Likelihood Ratio test to compare the model(s) to the base
model (if the provided model has more parameters than the base model).
\item Computes McFadden's Pseudo R^2.
}

The Likelihood-Ratio test is computed as \deqn{LR = -2 (LL_{base} - LL_{model})}.
The test is chi-squared with degrees of freedom
\deqn{dof=N_{model \ params}-N_{base \ params}}.
The AIC is calculated as \deqn{AIC = -2 \cdot LL + 2 \cdot nparam}, and the
BIC is calculated as \deqn{BIC = -2 \cdot LL + nparam \cdot \log(n)}.
}
\examples{

# Comparing the NBP model with the NB2 model and base Poisson model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0)

nbp.base <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 +
                    ShouldWidth04 + AADTover10k,
                    data=washington_roads, family = 'NBP', method = 'NM',
                    max.iters=3000)
nb2.base <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 +
                    ShouldWidth04 + AADTover10k,
                    data=washington_roads, family = 'NB2', method = 'NM',
                    max.iters=3000)                    
                   
regCompTest(nbp.base, washington_roads, basemodel="Poisson", 
model2=nb2.base, print=TRUE)

}
