% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renb.R
\name{renb}
\alias{renb}
\title{Estimate a Random Effects Negative Binomial regression model}
\usage{
renb(
  formula,
  group_var,
  data,
  method = "NM",
  max.iters = 1000,
  print.level = 0,
  bootstraps = NULL,
  offset = NULL
)
}
\arguments{
\item{formula}{an R formula.}

\item{group_var}{the grouping variable(s) for the random effects (e.g.,
individual ID or other panel ID variables).}

\item{data}{a dataframe that has all of the variables in the \code{formula}.}

\item{method}{a method to use for optimization in the maximum likelihood
estimation. For options, see \code{\link[maxLik]{maxLik}}. Note that "BHHH"
is not available for this function due to the implementation for the random
effects.}

\item{max.iters}{the maximum number of iterations to allow the optimization
method to perform.}

\item{print.level}{Integer specifying the verbosity of output during
optimization.}

\item{bootstraps}{Optional integer specifying the number of bootstrap samples
to be used for estimating standard errors. If not specified, no
bootstrapping is performed.}

\item{offset}{an optional offset term provided as a string.}
}
\value{
An object of class \code{countreg} which is a list with the following components:
\itemize{
\item model: the fitted model object.
\item data: the data frame used to fit the model.
\item call: the matched call.
\item formula: the formula used to fit the model.
}
}
\description{
Estimate a Random Effects Negative Binomial regression model
}
\details{
This function estimates a random effects negative binomial (RENB) regression
model. This model is based on the NB-1 model. The PDF for the RENB is:
\deqn{f(y_{it}|\lambda_{it}, a, b) = 
  \frac{\Gamma(a+b) 
    \Gamma(a + \sum_{t = 1}^{n_i} \\lambda_{it}) 
    \Gamma(b + \sum_{t=1}^{n_i}y_{it})}
    {\Gamma(a) \Gamma(b) \Gamma(a + b + 
       \sum_{t=1}^{n_i}\lambda_{it} + \sum_{t=1}^{n_i}y_{it})} \prod_{t=1}^{n_i}
       \frac{\Gamma(\lambda_{it}+y_{it})}{\Gamma(\lambda_{it})\Gamma(y_{it})}}

Where \eqn{y_{it}} is the count outcome for individual \eqn{i} at time
\eqn{t}, and \eqn{\lambda_{it}} is the latent Poisson mean parameter for
individual \eqn{i} at time \eqn{t}. The parameters \eqn{a} and \eqn{b} are
the shape parameters for the beta distribution that is used to model the
random effects. The RENB model allows for overdispersion in the count data
and accounts for unobserved heterogeneity across individuals by including
random effects in the model. This formulation follows the approach described
in the paper by Hausman, Hall, and Griliches (1984) for modeling panel data
with random effects.

The marginal mean and marginal variance of the RENB model are given by:
\deqn{E[y_{it}] = \lambda_{it}\frac{b}{a-1}=\mu_it=exp(X_{it}\beta)}

\deqn{Var[y_{it}] = \frac{a+b-1}{a-2}\mu_{it} + \frac{a+b-1}{b(a-2)} \mu_{it}^2}

Thus, the formulation of the model estimated here allows the use of the
estimated coefficients to directly compute the marginal mean.

Note that the RENB model is a panel data model, and the \code{group_var}
argument must be specified to indicate the grouping variable(s) for the
random effects. The model is estimated using maximum likelihood estimation,
and the optimization is performed using the \code{\link[maxLik]{maxLik}}
package. The user can specify the optimization method and maximum iterations.
}
\examples{
\donttest{
## RENB Model
data("washington_roads")
washington_roads$AADTover10k <- 
  ifelse(washington_roads$AADT > 10000, 1, 0) # create a dummy variable
renb.mod <- renb(Animal ~ lnaadt + speed50 + ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                offset = "lnlength",
                                group_var="ID",
                                method="nm",
                                max.iters = 1000)
summary(renb.mod)
}

}
\references{
Hausman, Jerry A., Bronwyn H. Hall, and Zvi Griliches. "Econometric models
for count data with an application to the patents–R&D relationship."
Econometrica: Journal of the Econometric Society (1984): 909-938.
}
