% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repois.R
\name{repois}
\alias{repois}
\title{Estimate a Random Effects Poisson Panel Model}
\usage{
repois(
  formula,
  group_var,
  data,
  method = "NM",
  max.iters = 1000,
  print.level = 0,
  bootstraps = NULL,
  offset = NULL
)
}
\arguments{
\item{formula}{an R formula.}

\item{group_var}{the grouping variable(s) for the random effects (e.g.,
individual ID or other panel ID variables).}

\item{data}{a dataframe that has all of the variables in the \code{formula}.}

\item{method}{a method to use for optimization in the maximum likelihood
estimation. For options, see \code{\link[maxLik]{maxLik}}.}

\item{max.iters}{the maximum number of iterations to allow the optimization
method to perform.}

\item{print.level}{integer specifying the verbosity of output during
optimization.}

\item{bootstraps}{optional integer specifying the number of bootstrap samples
to be used for estimating standard errors. If not specified, no
bootstrapping is performed.}

\item{offset}{an optional offset term provided as a string.}
}
\value{
An object of class \code{countreg} which is a list with the following
components:
\itemize{
\item model: the fitted model object.
\item data: the data frame used to fit the model.
\item call: the matched call.
\item formula: the formula used to fit the model.
}
}
\description{
Estimate a random effects Poisson panel model for clustered count data.
The model is also known as a negative multinomial model.
}
\details{
This function estimates a random effects Poisson panel model. The likelihood
is integrated analytically over a gamma-distributed cluster effect with mean
1 and variance \code{alpha}. The model estimates the regression coefficients
and one heterogeneity parameter \code{ln(alpha)}.
}
\examples{
\donttest{
## RE Poisson Panel Model
data("washington_roads")
washington_roads$AADTover10k <-
  ifelse(washington_roads$AADT > 10000, 1, 0)

repois.mod <- repois(
  Animal ~ lnaadt + speed50 + ShouldWidth04 + AADTover10k,
  data = washington_roads,
  offset = "lnlength",
  group_var = "ID",
  method = "BFGS",
  max.iters = 1000
)

summary(repois.mod)
}

}
