% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinvgaus.R
\name{rinvgaus}
\alias{rinvgaus}
\title{Generate inverse-Gaussian random numbers}
\usage{
rinvgaus(n, mean = 1, shape = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{mean}{Mean parameter (> 0).}

\item{shape}{Shape parameter (> 0).}
}
\value{
A numeric vector of length n.
}
\description{
Generates random variables from an inverse-Gaussian distribution using
the Michael, Schucany, and Haas (1976) method.
}
\details{
Parameterization:
IG(mean = mu, shape = lambda)
}
\keyword{internal}
