% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcount.R
\name{rpcount}
\alias{rpcount}
\title{Random Parameters Negative Binomial Model}
\usage{
rpcount(
  formula,
  rpar_formula,
  data,
  family = "nb2",
  rpardists = NULL,
  ndraws = 1500,
  scrambled = FALSE,
  correlated = FALSE,
  panel = NULL,
  weights = NULL,
  offset = NULL,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0
)
}
\arguments{
\item{formula}{An R formula for the fixed part of the model.}

\item{rpar_formula}{A one-sided formula for the random parameters (e.g., `~ var1 + var2`).}

\item{data}{A data frame containing all model variables.}

\item{family}{The model distribution type (e.g., "NB1", "NB2", "NBP").}

\item{rpardists}{A named character vector specifying the distribution for each random
parameter. Options: "n" (normal), "ln" (log-normal), "t" (triangular), "u" (uniform), "g" (gamma).
Defaults to normal for all.}

\item{ndraws}{The number of Halton draws for simulation.}

\item{scrambled}{Logical. If `TRUE`, uses scrambled Halton draws.}

\item{correlated}{Logical. If `TRUE`, estimates correlated random parameters.
Only the normal distribution is supported for correlated parameters.}

\item{panel}{A character vector of variable names defining the panel structure.}

\item{weights}{A string specifying the variable name for frequency weights.}

\item{offset}{A character vector of variable names to be used as an offset.}

\item{method}{The optimization method for `maxLik`. See `?maxLik::maxLik`.}

\item{max.iters}{Maximum iterations for the optimizer.}

\item{start.vals}{An optional named vector of starting values.}

\item{print.level}{Verbosity level for the optimizer (0, 1, or 2).}
}
\description{
Estimates a random parameters count model for any of the distributions also used in `countreg`
}
\examples{
\donttest{
data("washington_roads")

## Uncorrelated Random Parameters Poisson-Lognormal Model
nb2.rp <- rpcount(
  Total_crashes ~ lnlength + lnaadt,
  rpar_formula = ~ -1 + speed50,
  data = washington_roads,
  ndraws = 50, # Using fewer draws for a quick example
  rpardists = c(speed50 = "n"),
  family = 'PLN',
  method = "bfgs"
)
summary(nb2.rp)

## Correlated Random Parameters NB1 Model
nb1.rp.corr <- rpcount(
  Total_crashes ~ lnlength,
  rpar_formula = ~ lnaadt + speed50,
  data = washington_roads,
  ndraws = 50,
  correlated = TRUE,
  family = 'nb1',
  method = "bfgs"
)
summary(nb1.rp.corr)
}
}
