% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpnb.predict.R
\name{rpnb.predict}
\alias{rpnb.predict}
\title{Function for generating predictions based on the random parameters negative binomial with multiple optional methods}
\usage{
rpnb.predict(model, data, method = "Exact")
}
\arguments{
\item{model}{a model object estimated using the \code{rpnb} function,}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula}. This can be the data used for estimating the model or another dataframe,}

\item{method}{the method to be used in generating the predictions (options include \code{Simulated}, \code{Exact}, or \code{Individual}).}
}
\description{
Function for generating predictions based on the random parameters negative binomial with multiple optional methods
}
\note{
the method option \code{Individual} requires that the outcome be observed for all observations. This is due to the use of Bayesian methods for computing the individual observation coefficients.
}
\examples{
\donttest{

## Random Parameters Negative Binomial model (NB-2)

data("washington_roads")

nb2.rp <- rpnb(Total_crashes ~ - 1 + lnlength + lnaadt,
               rpar_formula = ~ speed50,
               data = washington_roads,
               ndraws = 100,
               correlated = TRUE,
               form = 'nb2',
               method = "bfgs",
               print.level = 1)

## Exact Prediction
hist(rpnb.predict(nb2.rp, washington_roads))

## Simulated Prediction
hist(rpnb.predict(nb2.rp, washington_roads, method="Simulated"))

## Exact Prediction
hist(rpnb.predict(nb2.rp, washington_roads, method="Individual"))
}
}
