% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppLind.R
\name{rppLind}
\alias{rppLind}
\title{Function for estimating a random parameter Poisson-Lindley model}
\usage{
rppLind(
  formula,
  rpar_formula,
  data,
  rpardists = NULL,
  ndraws = 1500,
  scrambled = FALSE,
  correlated = FALSE,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0,
  weights = NULL,
  bootstraps = NULL
)
}
\arguments{
\item{formula}{an R formula.. This formula should specify the outcome and the independent variables that have fixed parameters.}

\item{rpar_formula}{a symbolic description of the model related specifically to the random parameters. This should not include an outcome variable. If the intercept is random, include it in this formula. If the intercept is fixed, include it in \code{formula} but not in \code{rpar_formula}. To remove the intercept, use \code{0 + vars} or \code{-1 + vars},}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula},}

\item{rpardists}{an optional named vector whose names are the random parameters and values the distribution. The distribution options include normal ("n"), lognormal ("ln"), triangular ("t"), uniform ("u"), and gamma ("g"). If this is not provided, normal distributions are used for all random coefficients,}

\item{ndraws}{the number of Halton draws to use for estimating the random parameters,}

\item{scrambled}{if the Halton draws should be scrambled or not. \code{scrambled = FALSE} results in standard Halton draws while \code{scrambled = TRUE} results in scrambled Halton draws,}

\item{correlated}{if the random parameters should be correlated (\code{correlated = FALSE} results in uncorrelated random coefficients, \code{correlated = TRUE} results in correlated random coefficients). If the random parameters are correlated, only the normal distribution is used for the random coefficients,}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform,}

\item{start.vals}{an optional vector of starting values for the regression coefficients}

\item{print.level}{Integer specifying the verbosity of output during optimization.}

\item{weights}{the name of the weighting variable. This is an optional 
parameter for weighted regression.}

\item{bootstraps}{Optional integer specifying the number of bootstrap samples to be used
for estimating standard errors. If not specified, no bootstrapping is performed.}
}
\description{
Function for estimating a random parameter Poisson-Lindley model
}
\examples{
\dontrun{
## Random Parameters Poisson-Lindley
data("washington_roads")
plind.rp <- rppLind(Animal ~ lnlength + lnaadt,
                             rpar_formula = ~ -1 + speed50,
                             data = washington_roads,
                             ndraws = 10,
                             correlated = FALSE,
                             rpardists = c(speed50="n"),
                             method = "nm",
                             print.level = 0)

summary(plind.rp)}
}
