% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppois.R
\name{rppois}
\alias{rppois}
\title{Function for estimating a random parameters Poisson regression model}
\usage{
rppois(
  formula,
  rpar_formula,
  data,
  rpardists = NULL,
  ndraws = 1500,
  scrambled = FALSE,
  correlated = FALSE,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0
)
}
\arguments{
\item{formula}{An R formula specifying the outcome and the independent variables with fixed parameters.}

\item{rpar_formula}{A symbolic description of the model related to the random parameters. This should not include an outcome variable. If the intercept is random, include it in this formula. If the intercept is fixed, include it in \code{formula} but not in \code{rpar_formula}. To remove the intercept, use \code{0 + vars} or \code{-1 + vars}.}

\item{data}{A dataframe containing all of the variables in the \code{formula} and \code{rpar_formula}.}

\item{rpardists}{An optional named vector specifying the distributions of the random parameters. Options include normal ("n"), lognormal ("ln"), triangular ("t"), uniform ("u"), and gamma ("g"). If not provided, normal distributions are used for all random coefficients.}

\item{ndraws}{The number of Halton draws to use for estimating the random parameters.}

\item{scrambled}{Logical, whether the Halton draws should be scrambled. \code{FALSE} results in standard Halton draws, while \code{TRUE} results in scrambled Halton draws.}

\item{correlated}{Logical, whether the random parameters should be correlated. \code{FALSE} results in uncorrelated random coefficients, \code{TRUE} results in correlated random coefficients. If \code{TRUE}, only the normal distribution is used for the random coefficients.}

\item{method}{The optimization method to use for maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}}.}

\item{max.iters}{The maximum number of iterations for the optimization method.}

\item{start.vals}{An optional vector of starting values for the regression coefficients.}

\item{print.level}{Determines the verbosity level for printing optimization details. \code{0} prints no information, \code{1} prints minimal information, \code{2} prints detailed information.}
}
\description{
This function estimates a Poisson regression model with random parameters.
}
\examples{
\donttest{
## Random Parameters Poisson Regression model
data("washington_roads")
rpp <- rppois(Total_crashes ~ -1 + lnlength + lnaadt,
              rpar_formula = ~ speed50,
              data = washington_roads,
              ndraws = 100,
              correlated = FALSE,
              rpardists = c(intercept = "u", speed50 = "t"),
              method = "bfgs",
              print.level = 2)
summary(rpp)

Generate predictions and create a CURE Plot
}
}
