% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppoisson.R
\name{rppoisson}
\alias{rppoisson}
\title{Random Parameters Poisson Model}
\usage{
rppoisson(
  formula,
  rpar_formula,
  data,
  rpardists = NULL,
  ndraws = 1500,
  scrambled = FALSE,
  correlated = FALSE,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0,
  bootstraps = NULL
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted, specifying the outcome
and the independent variables with fixed effects.}

\item{rpar_formula}{A symbolic description of the model for random parameters,
excluding the outcome variable. Use 0 + vars or -1 + vars to exclude the intercept.}

\item{data}{A dataframe containing all variables included in 'formula' and 'rpar_formula'.}

\item{rpardists}{An optional named vector specifying the distributions of the random parameters.
Possible values include normal ("n"), lognormal ("ln"), and others as needed.
Defaults to normal if not specified.}

\item{ndraws}{The number of Halton draws for estimating random parameters.}

\item{scrambled}{Boolean to determine if Halton sequences should be scrambled.}

\item{correlated}{Boolean indicating if random parameters are correlated.
If TRUE, only normal distributions are used for random parameters.}

\item{method}{Optimization method to be used for maximum likelihood estimation.
See `maxLik` documentation for options.}

\item{max.iters}{Maximum number of iterations for the optimization method.}

\item{start.vals}{Optional vector of starting values for the optimization.}

\item{print.level}{Integer specifying the verbosity of output during optimization.}

\item{bootstraps}{Optional integer specifying the number of bootstrap samples to be used
for estimating standard errors.}
}
\value{
An object of class 'rppoisson' containing the model's estimated coefficients, the variance-covariance matrix of the random parameters (if correlated),
        standard errors, and other diagnostic measures.
}
\description{
Estimates a Poisson regression model with the possibility of random parameters. This function
allows for the specification of parameters that can vary randomly across observations,
modeled through the inclusion of Halton draws.
}
\examples{
data("washington_roads")
poisson_rp <- rppoisson(Total_crashes ~ lnlength + lnaadt,
                        rpar_formula = ~ speed50,
                        data = washington_roads,
                        ndraws = 10,
                        correlated = FALSE)
print(summary(poisson_rp))

}
