% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sichel.R
\name{sichel}
\alias{sichel}
\title{Function for estimating a Sichel regression model}
\usage{
sichel(formula, data, method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula}.}

\item{method}{a method to use for optimization in the maximum likelihood
estimation. For options, see \code{\link[maxLik]{maxLik}}.}

\item{max.iters}{the maximum number of iterations to allow the optimization
method to perform.}
}
\description{
Function for estimating a Sichel regression model
}
\details{
The Sichel regression model is based on the Sichel Distribution .

The expected value of the distribution in the regression utilizes a log-link
function. Thus, the mean is: \deqn{\mu=e^{X\beta}}

The variance is:
\deqn{
   \sigma^2= 
     \mu + 
     \left(\frac{2\sigma(\gamma+1)}{c} + \frac{1}{c^2} - 1\right) \mu^2}

The parameter \eqn{\sigma} is estimated as the natural logarithm transformed
value, ln(sigma), to ensure that \eqn{\sigma>0}.
}
\examples{
\donttest{

# Sichel Model
data("washington_roads")

sichel.mod <- sichel(Total_crashes ~ lnaadt + lnlength,
                                data=washington_roads,
                                method="NM",
                                max.iters = 1000)
summary(sichel.mod)}
}
\seealso{
\code{\link{dsichel}()} for more information on the distribution.
}
