% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rpnb.R
\name{summary.maxLik}
\alias{summary.maxLik}
\title{Custom summary method for maxLik objects}
\usage{
\method{summary}{maxLik}(
  object,
  confint_level = 0.95,
  digits = max(3, getOption("digits") - 3),
  ...
)
}
\arguments{
\item{object}{An object of class maxLik.}

\item{confint_level}{A numeric value between 0 and 1 indicating the confidence level for confidence intervals. Default is 0.95.}

\item{digits}{Number of digits to round to.}
}
\description{
Custom summary method for maxLik objects
}
