% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rpnb.R
\name{summary.rpnb}
\alias{summary.rpnb}
\title{Summary method for random parameter negative binomial model}
\usage{
\method{summary}{rpnb}(
  object,
  confint_level = 0.95,
  digits = max(3, getOption("digits") - 3),
  ...
)
}
\arguments{
\item{object}{a rpnb model, as produced by the rpnb function.}

\item{confint_level}{A numeric value between 0 and 1 indicating the confidence level for confidence intervals. Default is 0.95.}

\item{digits}{Number of digits to round to.}
}
\description{
Summary method for random parameter negative binomial model
}
