% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tidy.flexCountReg}
\alias{tidy.flexCountReg}
\title{Tidy a flexCountReg object}
\usage{
\method{tidy}{flexCountReg}(x, ...)
}
\arguments{
\item{x}{An object of class flexCountReg}

\item{...}{Additional arguments}
}
\description{
Tidy a flexCountReg object
}
