% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_to_use_predict.R
\name{veryCoolMethod}
\alias{veryCoolMethod}
\title{Create a function that returns an object of class veryCoolMethod}
\usage{
veryCoolMethod(mydata, ...)
}
\arguments{
\item{mydata}{Dataframe with at least two columns}

\item{...}{Additional arguments}
}
\value{
A list
}
\description{
Create a function that returns an object of class veryCoolMethod
}
\examples{
sample_object <- veryCoolMethod(cars)
sample_object
predict(sample_object)
}
