% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countreg.rp.R
\name{countreg.rp}
\alias{countreg.rp}
\title{Count regression models}
\usage{
countreg.rp(
  formula,
  data,
  family = "NB2",
  offset = NULL,
  weights = NULL,
  verbose = FALSE,
  dis_param_formula_1 = NULL,
  dis_param_formula_2 = NULL,
  underreport_formula = NULL,
  underreport_family = "logit",
  ndraws = 1500,
  method = "NM",
  max.iters = 1000,
  start.vals = NULL,
  stderr = "normal",
  bootstraps = NULL
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{family}{the name of the distribution/model type to estimate. The 
default "NB2" is the standard negative binomial distribution with a 
log link. other options are listed below.}

\item{offset}{the name of a variable in the data frame that should be used 
as an offset (i.e., included but forced to have a coefficient of 1).}

\item{weights}{the name of a variable in the data frame that should be used
as a frequency weight.}

\item{verbose}{an optional parameter. If `TRUE`, the function will print out 
the progress of the model fitting. Default is `FALSE`.}

\item{dis_param_formula_1}{a symbolic description of the model for the 
natural log of the dispersion parameter or first parameter of the 
count distribution used. Further details are provided below.}

\item{dis_param_formula_2}{a symbolic description of the model for the second
parameter of the count distribution used. Further details are provided 
below.}

\item{underreport_formula}{an optional formula to estimate the underreporting 
for any of the count model options. The underreporting is estimated as 
a function (logit or probit) of the predictors in the model. For the 
model to be tractable, the independent variables cannot be the exact 
same as the count model. The default is `NULL`.}

\item{underreport_family}{the name of the distribution/model type to estimate 
the underreporting portion of the model when `underreport_formula` is 
specified. The default is "logit" for a binary logistic regression 
model. The other option is "probit" for a probit model.}

\item{ndraws}{The number of Halton draws for integrating the distribution 
being compounded with the Poisson distribution when there is not a 
closed-form solution. This is also used in estimating the random 
parameters for random parameter models. Default is 1500. It is 
recommended to test different numbers of draws to determine if the 
model is stable (i.e., doesn't change or has minimal change as the 
number of draws changes within a reasonable range).}

\item{method}{Optimization method to be used for maximum likelihood 
estimation. See `maxLik` documentation for options. The default is 
"NM" for the Nelder-Mead method.}

\item{max.iters}{Maximum number of iterations for the optimization method.}

\item{start.vals}{Optional vector of starting values for the optimization.}

\item{stderr}{Type of standard errors to use. The default is "normal". Other 
options include "boot" for bootstrapped standard errors, or "robust" 
for robust standard errors.}

\item{bootstraps}{Optional integer specifying the number of bootstrap samples 
to be used for estimating standard errors when `stderr`= "boot". Note
that this currently does not work when an offset variable is used.}

\item{rpar_obs_formula}{an R formula with the independent variables that are 
observation-specific random parameters. This should not include an 
outcome variable. If the intercept is random (and varies over all 
observations), include it in this formula. If the intercept is fixed, 
include it in `formula` and not in `rpar_obs_formula`. To remove the 
intercept, use \code{0 + vars} or \code{-1 + vars}.}

\item{rpar_panel_formula}{an R formula with the independent variables that
are panel-specific random parameters. This should not include an 
outcome variable. If the intercept is random (and varies across the 
panels), include it in this formula. Only use this if the `panelID` is 
also specified.}

\item{rpar_underreport_formula}{an R formula with the independent variables
that are random parameters in the underreporting model. This should  
not include an outcome variable. If the intercept for the 
underreporting model is random include it in this formula and remove 
it from `underreport_formula`. Only use this if the 
`underreport_formula` is also specified.}

\item{rpar_obs_dists}{an optional named vector whose names are the random 
parameters and distribution to be used for estimating the random 
parameters for the variables in `rpar_obs_formula`. The distribution 
options include normal ("n"), lognormal ("ln"), triangular ("t"), 
uniform ("u"), and gamma ("g"). If this is not provided, normal 
distributions are used for all random coefficients in 
`rpar_obs_formula`. If `correlated`=`TRUE`, the distributions are  
set to the normal distribution, regardless of the distributions  
specified with `rpar_obs_dists`.}

\item{rpar_panel_dists}{an optional named vector whose names are the random 
parameters and distribution to be used for estimating the random 
parameters for the variables in `rpar_panel_formula`. The distribution 
options include normal ("n"), lognormal ("ln"), triangular ("t"), 
uniform ("u"), and gamma ("g"). If this is not provided, normal 
distributions are used for all random coefficients in 
`rpar_panel_formula`. If `correlated`=`TRUE`, the distributions are 
set to the normal distribution, regardless of the distributions  
specified with `rpar_panel_dists`.}

\item{rpar_underreport_dists}{an optional named vector whose names are the
random parameters and distribution to be used for estimating the  
random parameters for the variables in `rpar_underreport_formula`. The  
distribution options include normal ("n"), lognormal ("ln"), 
triangular ("t"),  uniform ("u"), and gamma ("g"). If this is not 
provided, normal distributions are used for all random coefficients in 
`rpar_underreport_formula`. If `correlated`=`TRUE`, the distributions 
are set to the normal distribution, regardless of the distributions  
specified with `rpar_underreport_dists`.}

\item{correlated}{an optional parameter. When a random parameter model is 
specified, a `TRUE` value indicates that the random parameters should
 be correlated. This uses correlated normal distributions with the 
 correlation captured using a Cholesky decomposition matrix. Default 
 is `FALSE`.}

\item{scrambled}{TRUE or FALSE. Indicates if the Halton draws should be 
scrambled. This removes correlation between Halton draws with larger 
base prime numbers. Default is TRUE.}

\item{panelID}{an optional string or vector of strings with the name(s) of 
variables in the data that identify the panels. This is only 
applicable to random effects and random parameter models. Default is 
`NULL`.}
}
\value{
An object of class `countreg` which is a list with the following components:
\itemize{
 \item model: the fitted model object.
 \item data: the data frame used to fit the model.
 \item call: the matched call.
 \item formula: the formula used to fit the model.
}
}
\description{
The purpose of this function is to estimate count regression models using 
maximum likelihood estimation (MLE) or Maximum Simulated Likelihood 
Estimation (MSLE). The function can estimate the following models:
\itemize{
 \item Poisson (Poisson)
 \item Negative Binomial 1 (NB1)
 \item Negative Binomial 2 (NB2)
 \item Negative Binomial P (NBP)
 \item Poisson-Lognormal (PLN)
 \item Poisson-Generalized-Exponential (PGE)
 \item Poisson-Inverse-Gaussian Type 1 (PIG1)
 \item Poisson-Inverse-Gaussian Type 2 (PIG2)
 \item Poisson-Lindley (PL)
 \item Poisson-Lindley-Gamma (PLG), also known as the Negative 
       Binomial-Lindley (NBL)
 \item Poisson-Lindley-Lognormal (PLL)
 \item Poisson-Weibull (PW)
 \item Sichel (SI)
 \item Generalized Waring (GW)
 \item Conway-Maxwell-Poisson (COM)
}
}
\details{
For the `family` argument, the following options are available:
\itemize{
 \item "Poisson" for the Poisson distribution with a log link.
 \item "NB1" for Negative Binomial 1 distribution with a log link.
 \item "NB2" for Negative Binomial 2 distribution with a log link (i.e., the 
       standard negative binomial model).
 \item "NBP" for Negative Binomial P distribution with a log link.
 \item "PLN" for Poisson-Lognormal distribution with a log link.
 \item "PGE" for Poisson-Generalized-Exponential distribution with a log 
       link.
 \item "PIG1" for Poisson-Inverse-Gaussian Type-1 distribution with a log link.
 \item "PIG2" for Poisson-Inverse-Gaussian Type-2 distribution with a log link.
 \item "PL" for Poisson-Lindley distribution with a log link.
 \item "PLG" for Poisson-Lindley-Gamma distribution with a log link.
 \item "PLL" for Poisson-Lindley-Lognormal distribution with a log link.
 \item "PW" for Poisson-Weibull distribution with a log link.
 \item "SI" for Sichel distribution with a log link.
 \item "GW" for Generalized Waring distribution with a log link.
 \item "COM" for Conway-Maxwell-Poisson (COM) distribution with a log link.
 }
 
 The `dis_param_formula_1` and `dis_param_formula_2` parameters are used to 
 estimate the dispersion parameter or other parameters of the count 
 distribution used. This leads to the distributions parameters being 
 functions rather than constants in the model. For example, 
 if the user wants to estimate the overdispersion parameter of the Negative 
 Binomial 2 distribution as a function of the variable `x1` and `x2`, 
 the user would specify `dis_param_formula_1 = ~ x1 + x2`. In the case of the 
 Negative Binomial distributions, the model is known as a Generalized 
 Negative Minomial model when the overdispersion parmater is specified as a 
 function.
 
 The function linking the distribution parameters to the predictors is:
 \deqn{Param = \exp((Intercept) + \sum \beta X)}
 
 The parameters for the different models are as follows:
 
 For `dis_param_formula_1`, the models are for the parameters:
 \itemize{
 \item Not Applicable for the Poisson model.
 \item \eqn{ln(\alpha)} for the Negative Binomial 1 model.
 \item \eqn{ln(\alpha)} for the Negative Binomial 2 model.
 \item \eqn{ln(\alpha)} for the Negative Binomial P model.
 \item \eqn{ln(\sigma)} for the Poisson-Lognormal model.
 \item shape parameter for the Poisson-Generalized-Exponential model.
 \item \eqn{ln(\eta)} for the Poisson-Inverse-Gaussian model.
 \item \eqn{ln(\theta)} for the Poisson-Lindley model.
 \item \eqn{ln(\theta)} for the Poisson-Lindley-Gamma model.
 \item \eqn{ln(\theta)} for the Poisson-Lindley-Lognormal model.
 \item \eqn{ln(\alpha)} for the Poisson-Weibull model.
 \item \eqn{\gamma} for the Sichel model.
 \item \eqn{k} for the Generalized Waring model.
 \item \eqn{ln(\nu)} for the Conway-Maxwell-Poisson model.
 }
 
 For `dis_param_formula_2`, the models are for the parameters:
 \itemize{
 \item Not Applicable for the Poisson model.
 \item Not Applicable for the Negative Binomial 1 model.
 \item Not Applicable for the Negative Binomial 2 model.
 \item p for the Negative Binomial P model.
 \item Not Applicable for the Poisson-Lognormal model.
 \item scale parameter for the Poisson-Generalized-Exponential model.
 \item Not Applicable for the Poisson-Inverse-Gaussian model.
 \item Not Applicable for the Poisson-Lindley model.
 \item \eqn{ln(\alpha)} for the Poisson-Lindley-Gamma model.
 \item \eqn{ln(\sigma)} for the Poisson-Lindley-Lognormal model.
 \item \eqn{ln(\sigma)} for the Poisson-Weibull model.
 \item \eqn{ln(\sigma)} for the Sichel model.
 \item \eqn{ln(\rho)} for the Generalized Waring model.
 \item Not Applicable for the Conway-Maxwell-Poisson model.
 }
 
 The `ndraws` parameter is used to estimate the distribution when there is 
 not a closed-form solution. This uses Halton draws to integrate the 
 distribution being compounded with the Poisson distribution. The default is 
 1500. The models this is applicable for include:
 \itemize{
 \item Poisson-Lognormal
 \item Poisson-Generalized-Exponential
 \item Poisson-Lindley-Gamma (more efficient and stable than using 
       hypergeometric functions)
 \item Poisson-Lindley-Lognormal
 \item Poisson-Weibull
 }
 The `ndraws` parameter is also used for the estimation of random parameters. 
 When there are 4 or more random parameters, it is recommended to use 
 `scrambeled` = TRUE to use scrambled Halton draws (which removes correlation 
 that is inherent between Halton draws with higher base prime numbers)
}
\section{Random Parameters Model Details}{

For random parameters, the model captures heterogeneity of effects and other 
unobserved heterogeneity correlated with specified parameters. These can 
accommodate panel (i.e., longitudinal) model specifications. When the model 
is based on panel data, random parameters can be allowed to vary across 
individuals (i.e., each individual has a constant coefficient/effect related 
to that variable but the coefficient varies across individuals) or within 
individuals (where every observation, regardless of the individual, has 
different coefficient values). The differences for this are captured using 
the model parameters `rpar_panel_formula` for the random parameters that vary 
across individuals (only applicable if the panel ID is also specified) and 
`rpar_obs_formula` for the random parameters that vary across all 
observations (i.e., within individuals). 

Additionally, random parameters for the underreporting model (if specified) 
can also be estimated using `rpar_underreport_formula`. When this is used 
with a panel model specification, the random parameters for underreporting 
are treated as across-individual random parameters. This increases the 
tractability of the model. 

For the random parameters, they can be correlated or uncorrelated. If they 
correlated (specified with `correlated`=`TRUE`), the random parameters all 
follow correlated normal distributions. If they are not correlated, the 
distributions can be specified for the variables in each of the random 
parameters formulas using `rpar_obs_dists` for `rpar_obs_formula`, 
`rpar_panel_dists` for `rpar_panel_formula`, and `rpar_underreport_dists` 
for `rpar_underreport_formula`. Options for the random parameter 
distributions include:
 \itemize{
 \item "n" for the normal distribution
 \item "ln" for the lognormal distribution
 \item "t" for the symmetrical triangular distribution 
 \item "u" for the uniform distribution
 \item "g" for the gamma distribution
 }
The normal distribution is most commonly used. The lognormal and gamma 
distributions are useful for coefficients that have a theoretical reason that 
the coefficients should always be positive or always be negative. The uniform 
distribution often works well for indicator variables.

To specify the distributions, provide a named vector of coefficient names 
with the associated distributions. For instance, if:

`rpar_obs_formula` ~ 1 + x1 + x2 

And the distributions desired are uniform for the intercept, gamma for x1, 
and normal for x2, the distributions would be specified as 

`rpar_obs_dists` = c(intercept="u", x1="g", x2="n")

**Special Case Warning**
When the intercept is included in `rpar_obs_formula` and is normally 
distributed and the `family`="PL", the results will be intractable. The 
Poisson-Lognormal model is equivalent to a random parameters Poisson model 
that has a normally distributed random intercept that varies across all 
observations.
}

\examples{
\dontrun{
# Load the Washington data
data("washington_roads")
washington_roads$AADT10kplus <- ifelse(washington_roads$AADT > 10000, 1, 0)
# Estimate an NB2 model with a dispersion parameter as a function of the 
# variable `speed50` (i.e., generalized NB2), verbose output, and use the 
# BFGS optimization method
nb2 <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 + AADT10kplus,
               data = washington_roads, family = "NB2",
               dis_param_formula_1 = ~ speed50, verbose = TRUE, method='BFGS')
summary(nb2)

# Estimate a Poisson-Lognormal model (a low number of draws is used to speed 
# up the estimation for examples - not recommended in practice)
pln <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 + AADT10kplus,
              data = washington_roads, family = "PLN", ndraws=10)
summary(pln)  

# Estimate an Poisson-Lognormal with underreporting (probit)
plogn_underreport <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 + AADT10kplus,
              data = washington_roads, family = "NB2",
              underreport_formula = ~ speed50 + AADT10kplus, underreport_family = "probit")
summary(plogn_underreport)

# Estimate a Conwa-Maxwell-Poisson model
com_model <- countreg(Total_crashes ~ lnaadt + lnlength + speed50 + AADT10kplus,
              data = washington_roads, family = "COM")
summary(com_model)
}

}
\references{
Greene, W. (2008). Functional forms for the negative binomial model for count data. Economics Letters, 99(3), 585-590.

Pararai, M., Famoye, F., & Lee, C. (2006). Generalized Poisson regression model for underreported counts. Advances and applications in Statistics, 6(3), 305-322.

Pararai, M., Famoye, F., & Lee, C. (2010). Generalized poisson-poisson mixture model for misreported counts with an application to smoking data. Journal of Data Science, 8(4), 607-617.

Rigby, R. A., Stasinopoulos, D. M., & Akantziliotou, C. (2008). A framework for modelling overdispersed count data, including the Poisson-shifted generalized inverse Gaussian distribution. Computational Statistics & Data Analysis, 53(2), 381-393.

Wood, J.S., Eric T. Donnell, & Christopher J. Fariss. "A method to account for and estimate underreporting in crash frequency research." Accident Analysis & Prevention 95 (2016): 57-66.

Zou, Y., Lord, D., & Zhang, Y. (2012). Analyzing highly dispersed crash data using the Sichel generalized additive models for location, scale and shape.
}
