% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generalized-Waring.R
\name{Generalized-Waring}
\alias{Generalized-Waring}
\alias{dgwar}
\alias{pgwar}
\alias{qgwar}
\alias{rgwar}
\title{Generalized Waring Distribution}
\usage{
dgwar(y, mu, k, rho, log = FALSE)

pgwar(q, mu, k, rho, lower.tail = TRUE, log.p = FALSE)

qgwar(p, mu, k, rho)

rgwar(n, mu, k, rho)
}
\arguments{
\item{y}{non-negative integer vector of count outcomes.}

\item{mu}{positive numeric vector of distribution means.}

\item{k}{positive numeric vector of shape parameters.}

\item{rho}{numeric vector of shape parameters greater than 1. A finite
variance requires \eqn{\rho > 2}.}

\item{log}{logical; if TRUE, log probabilities are returned.}

\item{q}{non-negative integer vector of quantiles.}

\item{lower.tail}{logical; if TRUE, probabilities are
\eqn{P[X \leq q]}; otherwise, \eqn{P[X > q]}.}

\item{log.p}{logical; if TRUE, probabilities are returned as log
probabilities.}

\item{p}{numeric vector of probabilities.}

\item{n}{integer number of random numbers to generate.}
}
\value{
\code{dgwar} gives the probability mass function, \code{pgwar} gives the
distribution function, \code{qgwar} gives the quantile function, and
\code{rgwar} generates random deviates.

The length of the result is determined by \code{n} for \code{rgwar} and is
the maximum of the lengths of the numerical arguments for the other
functions.
}
\description{
These functions provide the probability mass function, distribution
function, quantile function, and random number generation for the
Generalized Waring distribution.
}
\details{
The Generalized Waring distribution is a three-parameter count distribution
used to model overdispersed count data.

\code{dgwar} computes the probability mass function (PMF) of the
Generalized Waring distribution.

\code{pgwar} computes the CDF of the Generalized Waring distribution.

\code{qgwar} computes the quantile function of the Generalized Waring
distribution.

\code{rgwar} generates random numbers from the Generalized Waring
distribution.

The probability mass function for the Generalized Waring distribution is:
\deqn{
f(y \mid a_x, k, \rho) =
\frac{
  \Gamma(a_x + \rho)\Gamma(k + \rho)
  (a_x)_y(k)_y
}{
  y!\Gamma(\rho)\Gamma(a_x + k + \rho)
  (a_x + k + \rho)_y
}
}
where
\eqn{(\alpha)_r = \frac{\Gamma(\alpha+r)}{\Gamma(\alpha)}},
and \eqn{a_x > 0}, \eqn{k > 0}, and \eqn{\rho > 0}.

When \eqn{\rho > 1}, the mean is:
\deqn{
E[Y] = \frac{a_x k}{\rho - 1}.
}

Therefore, the distribution can be parameterized in terms of its mean using:
\deqn{
a_x = \frac{\mu(\rho - 1)}{k}.
}

For a regression model:
\deqn{
\mu = \exp(X\beta).
}

When \eqn{\rho > 2}, the variance under this mean parameterization is:
\deqn{
\mathrm{Var}(Y) =
\frac{\mu(k+\mu)(k+\rho-1)}{k(\rho-2)}.
}
}
\examples{
dgwar(0, mu = 1, k = 2, rho = 3)
pgwar(c(0, 1, 2, 3), mu = 1, k = 2, rho = 3)
qgwar(0.8, mu = 1, k = 2, rho = 3)
rgwar(10, mu = 1, k = 2, rho = 3)

}
