% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generalized-Waring.R
\name{Generalized-Waring}
\alias{Generalized-Waring}
\alias{dgwar}
\alias{pgwar}
\alias{qgwar}
\alias{rgwar}
\title{Generalized Waring Distribution}
\usage{
dgwar(y, mu, alpha, rho, log = FALSE)

pgwar(q, mu, alpha, rho, lower.tail = TRUE, log.p = FALSE)

qgwar(p, mu, alpha, rho)

rgwar(n, mu, alpha, rho)
}
\arguments{
\item{y}{non-negative integer vector of count outcomes.}

\item{mu}{numeric vector of means of the distribution.}

\item{alpha}{non-negative numeric parameter of the distribution.}

\item{rho}{non-negative numeric parameter of the distribution.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{non-negative integer vector of quantiles.}

\item{lower.tail}{logical; if TRUE, probabilities p are \eqn{P[X\leq x]} otherwise, \eqn{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{numeric vector of probabilities.}

\item{n}{integer number of random numbers to generate.}
}
\description{
These functions provide density, distribution function, quantile function, and random number generation for the Generalized Waring Distribution.
}
\details{
The Generalized Waring distribution is a 3-parameter count distribution that is used to model overdispersed count data.


\code{dwaring} computes the density (PMF) of the Generalized Waring Distribution.

\code{pwaring} computes the CDF of the Generalized Waring Distribution.

\code{qwaring} computes the quantile function of the Generalized Waring Distribution.

\code{rwaring} generates random numbers from the Generalized Waring Distribution.

The Probability Mass Function (PMF) for the Generalized Waring (GW) distribution is:
\deqn{PMF=\frac{\Gamma(\alpha+\rho)}{\Gamma(\alpha)\Gamma(\rho)}
       \left(\frac{\alpha}{\mu(\rho-1)}\right)^{\alpha-1}
       \left(1+\frac{\alpha}{\mu(\rho-1)}\right)^{-(\alpha+\rho)}}

Where \eqn{\alpha} and \eqn{\rho} are distribution parameters with the constraints that \eqn{\alpha \geq 0} and \eqn{\rho > 1}, \eqn{\mu} is the mean of the distribution.
}
\examples{
dgwar(0, mu=1, alpha=2, rho=3)
pgwar(c(0,1,2,3), mu=1, alpha=2, rho=3)
qgwar(c(0.1, 0.5, 0.9), mu=1, alpha=2, rho=3)
rgwar(10, mu=1, alpha=2, rho=3)

}
