% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genpoisson1.R
\name{GeneralizedPoisson}
\alias{GeneralizedPoisson}
\alias{dgp1}
\alias{pgp1}
\alias{qgp1}
\alias{rgp1}
\title{Generalized Poisson Version 1 Distribution}
\usage{
dgp1(x, mu = 1, phi = 1, log = FALSE)

pgp1(q, mu = 1, phi = 1, lower.tail = TRUE, log.p = FALSE)

qgp1(p, mu = 1, phi = 1, lower.tail = TRUE, log.p = FALSE)

rgp1(n, mu = 1, phi = 1)
}
\arguments{
\item{x}{numeric value or a vector of values.}

\item{mu}{numeric value or vector of mean values for the distribution (the
values have to be greater than 0).}

\item{phi}{single value or vector of values for the scale parameter of the
distribution (the values have to be greater than -1).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{quantile or a vector of quantiles.}

\item{lower.tail}{logical; if TRUE, probabilities p are \eqn{P[X\leq x]}
otherwise, \eqn{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{probability or a vector of probabilities.}

\item{n}{the number of random numbers to generate.}
}
\value{
dgp1 gives the density, pgp1 gives the distribution
function, qgp1 gives the quantile function, and rgp1 generates random
deviates.

The length of the result is determined by n for rgp1, and is the
maximum of the lengths of the numerical arguments for the other functions.
}
\description{
These functions provide the density function, distribution function,
quantile function, and random number generation for the
Generalized Poisson Version 1 (GP-1) Distribution
}
\details{
\code{dgp1} computes the density (PDF) of the Generalized Poisson
Distribution.

\code{pgp1} computes the CDF of the Generalized Poisson Distribution.

\code{qgp1} computes the quantile function of the
Generalized Poisson Distribution.

\code{rgp1} generates random numbers from the Generalized Poisson
Distribution.

The compound Probability Mass Function (PMF) for the Generalized Poisson
distribution, version 1, (GP-1) is:
\deqn{
f(y|\phi,\mu)=\frac{\mu(\mu+\phi y)^{y-1} exp\left(-\frac{\mu+\phi y}
{1+\phi}\right)}{(1+\phi)^y y!}
}

Where \eqn{\phi} is a scale parameter with the restriction that
\eqn{\mu>0} is the mean value, and \eqn{y} is a non-negative
integer. This formulation uses the mean directly.

The variance of the GP-1 distribution is:
\deqn{\sigma^2=(1+\phi)^2 \mu}

If \eqn{\phi>0}, the distribution is overdispersed. If \eqn{\phi=0}, the
distribution is equidispersed. If \eqn{\phi<0}, the distribution is
underdispersed.

Furthermore, \eqn{\phi>-1} is required for this distribution. When
\eqn{\phi<0}, there is also a maximum value of support for the integar value
\eqn{y}. This is \eqn{y_max=\left\lfloor -\frac{\mu}{\phi} \right\rfloor}.
}
\examples{
dgp1(1, mu=0.75, phi=-0.1)
pgp1(c(0,1,2,3,5,7,9,10), mu=0.75, phi=3)
qgp1(c(0.1,0.3,0.5,0.9,0.95), mu=0.75, phi=0.5)
rgp1(30, mu=0.75, phi=1.5)

}
\references{
Consul, PoC, and Felix Famoye. "Generalized Poisson regression model."
Communications in Statistics-Theory and Methods 21.1 (1992): 89-109.

Zamani, Hossein, and Noriszura Ismail. "Functional form for the generalized
Poisson regression model." Communications in Statistics-Theory and Methods
41.20 (2012): 3666-3675.
}
