% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genWaring.R
\name{genWaring}
\alias{genWaring}
\title{Function for estimating a Generalized Waring regression model}
\usage{
genWaring(formula, data, method = "BHHH", max.iters = 1000)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in  `formula`.}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}} documentation.}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform.}
}
\description{
Function for estimating a Generalized Waring regression model
}
\details{
## Generalized Waring Probability Mass Function, Mean, and Variance
The following are the versions of the PMF, mean, and variance used in this function. This is adjusted from the typical formulation by replacing parameter \code{k} with \eqn{\mu}
\deqn{PMF=\frac{\Gamma(\alpha+\rho)}{\Gamma(\alpha)\Gamma(\rho)}\left(\frac{\alpha}{\mu(\rho-1)}\right)^{\alpha-1}\left(1+\frac{\alpha}{\mu(\rho-1)}\right)^{-(\alpha+\rho)}}
\deqn{\mu=e^{X\beta}}
\deqn{\sigma^2=\mu+\left(\frac{\frac{\mu(\rho-1)}{\alpha}+1}{\rho-2}\right)\mu+\left(\frac{\frac{\mu(\rho-1)}{\alpha}+\rho+1}{\frac{\mu(\rho-1)}{\alpha}(\rho-2)}\right)\mu^2}
Where \eqn{\alpha} and \eqn{\rho} are distribution parameters with the constraints that \eqn{\alpha\geq 0} and \eqn{\rho\geq 0}, \eqn{X} is a matrix of predictors, and \eqn{\beta} is a vector of coefficients.
}
\examples{
\donttest{

## Generalized Waring Model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0) # create a dummy variable
genwaring.mod <- genWaring(Total_crashes ~ lnaadt + lnlength + speed50 +
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads)
summary(genwaring.mod)}
}
