% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbg.R
\name{nbg}
\alias{nbg}
\title{Function for estimating a variety of negative binomial models (NB-1, NB-2,
NB-P and generalized versions of each)}
\usage{
nbg(
  formula,
  data,
  form = "nb2",
  ln.alpha.formula = NULL,
  method = "BHHH",
  max.iters = 200
)
}
\arguments{
\item{formula}{an R formula.}

\item{data}{a dataframe that has all of the variables in the \code{formula}
and \code{rpar_formula}. This can be the data used for estimating the model
or another dataframe,}

\item{form}{the version of the negative binomial to estimate (\code{"nb2"}
estimates the NB-2, \code{"nb1"} estimates the NB-1, \code{"nbp"} estimates
the NB-P)}

\item{ln.alpha.formula}{an optional formula for using independent variables
to estimate the natural log of the overdispersion parameter (makes the
model a generalized negative binomial).}

\item{method}{a method to use for optimization in the maximum likelihood
estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization
method to perform.}
}
\description{
Function for estimating a variety of negative binomial models (NB-1, NB-2,
NB-P and generalized versions of each)
}
\details{
The NB-1, NB-2, and NB-P versions of the negative binomial distribution
}
\examples{
\donttest{

## NB-P model
data("washington_roads")
washington_roads$AADTover10k <- ifelse(washington_roads$AADT>10000,1,0)

nbp.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                    ShouldWidth04 + AADTover10k,
                    data=washington_roads, form = 'nbp', method = 'NM',
                    max.iters=3000)
summary(nbp.base)

## Generalized NB-P model

nbp.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                                ShouldWidth04 + AADTover10k,
                                data=washington_roads,
                                form = 'nbp',
                                method = 'NM',
                                max.iters=3000,
                                ln.alpha.formula = ~ 1+lnlength)
summary(nbp.overdispersion)

## NB-1 Model
nb1.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb1',
                        method = 'NM',
                        max.iters=3000)
summary(nb1.base)

## Generalize NB-1 Model
nb1.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb1',
                        method = 'NM',
                        max.iters=3000, ln.alpha.formula = ~ 1+lnlength)
summary(nb1.overdispersion)

## NB-2 Model
nb2.base <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb2',
                        method = 'NM',
                        max.iters=3000)
summary(nb2.base)

## Generalize NB-2 Model
nb2.overdispersion <- nbg(Total_crashes ~ lnaadt + lnlength + speed50 +
                        ShouldWidth04 + AADTover10k,
                        data=washington_roads, form = 'nb2',
                        method = 'NM',
                        max.iters=3000, ln.alpha.formula = ~ 1+lnlength)
summary(nb2.overdispersion)}
}
