% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plindGamma.R
\name{Negative-Binomial-Lindley}
\alias{Negative-Binomial-Lindley}
\alias{dplindGamma}
\alias{pplindGamma}
\alias{qplindGamma}
\alias{rplindGamma}
\title{Poisson-Lindley-Gamma (Negative Binomial-Lindley) Distribution}
\usage{
dplindGamma(
  x,
  mean = 1,
  theta = 1,
  alpha = 1,
  lambda = NULL,
  ndraws = 1500,
  log = FALSE
)

pplindGamma(
  q,
  mean = 1,
  theta = 1,
  lambda = NULL,
  alpha = 1,
  ndraws = 1500,
  lower.tail = TRUE,
  log.p = FALSE
)

qplindGamma(p, mean = 1, theta = 1, alpha = 1, ndraws = 1500, lambda = NULL)

rplindGamma(n, mean = 1, theta = 1, alpha = 1, ndraws = 1500, lambda = NULL)
}
\arguments{
\item{x}{numeric value or a vector of values.}

\item{mean}{numeric value or vector of mean values for the distribution (the values have to be greater than 0).}

\item{theta}{single value or vector of values for the theta parameter of the distribution (the values have to be greater than 0).}

\item{alpha}{single value or vector of values for the `alpha` parameter of the gamma distribution in the special case that the mean = 1 and the variance = `alpha` (the values for `alpha` have to be greater than 0).}

\item{lambda}{alternative parameterization (use instead of the mean); numeric value or vector of values for lambda parameter of the distribution (the values have to be greater than 0).}

\item{ndraws}{the number of Halton draws to use for the integration.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{quantile or a vector of quantiles.}

\item{lower.tail}{logical; if TRUE, probabilities p are \eqn{P[X\leq x]} otherwise, \eqn{P[X>x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{probability or a vector of probabilities.}

\item{n}{the number of random numbers to generate.}
}
\description{
These functions provide density, distribution function, quantile function, and random number generation for the Poisson-Lindley-Gamma (PLG) Distribution
}
\details{
The Poisson-Lindley-Gamma is a 3-parameter count distribution that captures high densities for small integer values and provides flexibility for heavier tails.


\code{dplindGamma} computes the density (PDF) of the Poisson-Lindley-Gamma Distribution.

\code{pplindGamma} computes the CDF of the Poisson-Lindley-Gamma Distribution.

\code{qplindGamma} computes the quantile function of the Poisson-Lindley-Gamma Distribution.

\code{rplindGamma} generates random numbers from the Poisson-Lindley-Gamma Distribution.

The compound Probability Mass Function (PMF) for the Poisson-Lindley-Gamma (PLG) distribution utilizes a special case
\deqn{f(y|\mu,\theta,\alpha)=\int_0^\infty \frac{\lambda^y x^y e^{-\lambda x}}{y!} \frac{x^{\frac{1}{\alpha}-1}e^{-\frac{x}{\alpha}}}{\Gamma\left(\frac{1}{\alpha}\right)\alpha^{\frac{1}{\alpha}}}dx}

Where \eqn{\theta} and \eqn{\lambda} are distribution parameters from the Poisson-Lindley distribution with the restrictions that \eqn{\theta>0} and \eqn{\lambda>0}, \eqn{\alpha} is a parameter for the gamma distribution with the restriction \eqn{\alpha>0}, and \eqn{y} is a non-negative integer.

The expected value of the distribution is:
\deqn{\mu=\frac{\lambda(\theta+2)}{\theta(\theta+1)}}

The default is to use the input mean value for the distribution. However, the lambda parameter can be used as an alternative to the mean value.

Halton draws are used to perform simulation over the gamma distribution to solve the integral.
}
\examples{
dplindGamma(0, mean=0.75, theta=7, alpha=2, ndraws=10)
pplindGamma(c(0,1,2,3,5,7,9,10), mean=0.75, theta=7, alpha=2, ndraws=10)
qplindGamma(c(0.1,0.3,0.5,0.9,0.95), lambda=4.67, theta=7, alpha=2,
              ndraws=10)
rplindGamma(3, lambda=4.67, theta=7, alpha=2, ndraws=10)

}
