% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisWeib.R
\name{PoissonWeibull}
\alias{PoissonWeibull}
\alias{dpoisweibull}
\alias{ppoisweibull}
\alias{qpoisweibull}
\alias{rpoisweibull}
\title{Poisson-Weibull Distribution Functions}
\usage{
dpoisweibull(
  x,
  alpha = NULL,
  beta = NULL,
  mean_value = NULL,
  sd_value = NULL,
  log = FALSE
)

ppoisweibull(
  q,
  alpha = NULL,
  beta = NULL,
  mean_value = NULL,
  sd_value = NULL,
  lower.tail = TRUE,
  log.p = FALSE
)

qpoisweibull(p, alpha = NULL, beta = NULL, mean_value = NULL, sd_value = NULL)

rpoisweibull(n, alpha = NULL, beta = NULL, mean_value = NULL, sd_value = NULL)
}
\arguments{
\item{x}{A numeric value or vector of values for which the PDF or CDF is calculated.}

\item{alpha}{Shape parameter of the Weibull distribution (optional if mean and sd are provided).}

\item{beta}{Scale parameter of the Weibull distribution (optional if mean and sd are provided).}

\item{mean_value}{Mean of the Weibull distribution (optional if alpha and beta are provided).}

\item{sd_value}{Standard deviation of the Weibull distribution (optional if alpha and beta are provided).}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{q}{Quantile or a vector of quantiles.}

\item{lower.tail}{Logical; if TRUE, probabilities are P[X <= x], otherwise P[X > x].}

\item{log.p}{Logical; if TRUE, probabilities p are given as log(p).}

\item{p}{A numeric value or vector of probabilities for the quantile function.}

\item{n}{The number of random samples to generate.}
}
\description{
These functions provide density, distribution function, quantile function, and random number 
generation for the Poisson-Weibull Distribution, which is specified either by its shape and scale 
parameters or by its mean and standard deviation.
}
\details{
The Poisson-Weibull distribution uses the Weibull distribution as a mixing distribution for a 
Poisson process. It is useful for modeling overdispersed count data. The density function 
(probability mass function) for the Poisson-Weibull distribution is given by:
\deqn{P(X = k) = \int_0^\infty e^{-\theta} \theta^k / k! f(\theta; \alpha, \beta) d\theta,}
where \eqn{f(\theta; \alpha, \beta)} is the PDF of the Weibull distribution.


\code{dpoisweibull} computes the density of the Poisson-Weibull distribution.

\code{ppoisweibull} computes the distribution function of the Poisson-Weibull distribution.

\code{qpoisweibull} computes the quantile function of the Poisson-Weibull distribution.

\code{rpoisweibull} generates random numbers following the Poisson-Weibull distribution.

The shape and scale parameters directly define the Weibull distribution, whereas the mean and 
standard deviation are used to compute these parameters indirectly.
}
\examples{
dpoisweibull(4, alpha = 1.5, beta = 0.5)
ppoisweibull(4, alpha = 1.5, beta = 0.5)
qpoisweibull(0.95, alpha = 1.5, beta = 0.5)
rpoisweibull(10, alpha = 1.5, beta = 0.5)

}
