% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_flexCountReg.R
\name{predict.flexCountReg}
\alias{predict.flexCountReg}
\title{Predictions for flexCountReg models}
\usage{
\method{predict}{flexCountReg}(object, ...)
}
\arguments{
\item{object}{a model object estimated using this R package.}

\item{...}{optional arguments passed to the function. These include `data` 
and `method`.}
}
\description{
Predictions for flexCountReg models
}
\note{
optional parameter `data`: a dataframe that has all of the variables in 
       the \code{formula} and \code{rpar_formula}.

optional parameter `method`: Only valid for random parameters models. 
       Options include \code{Simulated} (default) or \code{Individual}. 
       \code{Exact} is only available for the older `rpnb` class.
}
