% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_to_use_predict.R
\name{predict.veryCoolMethod}
\alias{predict.veryCoolMethod}
\title{Create a method for function print for class veryCoolMethod}
\usage{
\method{predict}{veryCoolMethod}(veryCoolMethodObject)
}
\arguments{
\item{veryCoolMethodObject}{}
}
\value{
Predicted valued for veryCoolMethod
}
\description{
Create a method for function print for class veryCoolMethod
}
\examples{
sample_object <- veryCoolMethod(cars)
sample_object
predict(sample_object)
}
