% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwiebreg.R
\name{pwiebreg}
\alias{pwiebreg}
\title{Poisson-Weibull Regression}
\usage{
pwiebreg(
  formula,
  alpha_formula = NULL,
  beta_formula = NULL,
  data,
  ndraws = 1500,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0,
  bootstraps = NULL
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted, specifying the outcome
and the independent variables with fixed effects.}

\item{alpha_formula}{A symbolic description of the model for the alpha parameter,
specifying the distribution parameter as a function of predictor variables.}

\item{beta_formula}{A symbolic description of the model for the beta parameter,
specifying the distribution parameter as a function of predictor variables.}

\item{data}{A dataframe containing all variables included in 'formula', 'alpha_formula', and 'beta_formula'.}

\item{ndraws}{The number of Halton draws for integrating the Weibull distribution.}

\item{method}{Optimization method to be used for maximum likelihood estimation.
See `maxLik` documentation for options.}

\item{max.iters}{Maximum number of iterations for the optimization method.}

\item{start.vals}{Optional vector of starting values for the optimization.}

\item{print.level}{Integer specifying the verbosity of output during optimization.}

\item{bootstraps}{Optional integer specifying the number of bootstrap samples to be used
for estimating standard errors.}
}
\description{
Estimates a Poisson-Weibull regression model.
}
\details{
For the Poisson-Weibull Regression model, the expected values is:
\deqn{E[Y]=\lambda\beta\Gamma\left(1+\frac{1}{\alpha}\right)}
Where \eqn{\lambda} is the mean of the Poisson distribution, \eqn{\alpha} is the shape parameter, and \eqn{\beta} is the scale parameter.

To ensure that the regression model predicts the mean value, the regression utilizes:
\deqn{\mu=\exp{X\gamma}=\lambda\beta\Gamma\left(1+\frac{1}{\alpha}\right)}
Where \eqn{X} is a matrix of independent variables and \eqn{\gamma} is a vector of coefficients. 

This leads to:
\deqn{\lambda=\frac{\mu}{\beta\Gamma\left(1+\frac{1}{\alpha}\right)}}

The variance for the Poisson-Weibull regression is:

\deqn{V[Y]=\mu+\left(\frac{\Gamma\left(1+\frac{2}{\alpha}\right)}{\Gamma\left(1+\frac{1}{\alpha}\right)^2}-1\right)\mu^2}
}
\examples{
data("washington_roads")
pw_rp <- pwiebreg(Total_crashes ~ lnlength + lnaadt,
                                 data = washington_roads,
                                 ndraws = 100)
print(summary(pw_rp))

}
