% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpnb.R
\name{rpnb}
\alias{rpnb}
\title{Function for estimating a random parameter negative binomial with the ability to specify if the NB-1, NB-2, or NB-P should be used}
\usage{
rpnb(
  formula,
  rpar_formula,
  data,
  form = "nb2",
  rpardists = NULL,
  ndraws = 1500,
  scrambled = FALSE,
  correlated = FALSE,
  panel = NULL,
  weights = NULL,
  offset = NULL,
  method = "BHHH",
  max.iters = 1000,
  start.vals = NULL,
  print.level = 0
)
}
\arguments{
\item{formula}{an R formula.. This formula should specify the outcome and the independent variables that have fixed parameters.}

\item{rpar_formula}{a symbolic description of the model related specifically to the random parameters. This should not include an outcome variable. If the intercept is random, include it in this formula. If the intercept is fixed, include it in \code{formula} but not in \code{rpar_formula}. To remove the intercept, use \code{0 + vars} or \code{-1 + vars},}

\item{data}{a dataframe that has all of the variables in the \code{formula} and \code{rpar_formula},}

\item{form}{the version of the negative binomial to estimate (\code{"nb2"} estimates the NB-2, \code{"nb1"} estimates the NB-1, \code{"nbp"} estimates the NB-P)}

\item{rpardists}{an optional named vector whose names are the random parameters and values the distribution. The distribution options include normal ("n"), lognormal ("ln"), triangular ("t"), uniform ("u"), and gamma ("g"). If this is not provided, normal distributions are used for all random coefficients,}

\item{ndraws}{the number of Halton draws to use for estimating the random parameters,}

\item{scrambled}{if the Halton draws should be scrambled or not. \code{scrambled = FALSE} results in standard Halton draws while \code{scrambled = TRUE} results in scrambled Halton draws,}

\item{correlated}{if the random parameters should be correlated (\code{correlated = FALSE} results in uncorrelated random coefficients, \code{correlated = TRUE} results in correlated random coefficients). If the random parameters are correlated, only the normal distribution is used for the random coefficients,}

\item{panel}{an optional variable or vector of variables that can be used to specify a panel structure in the data. If this is specified, the function will estimate the random parameters using a panel structure,}

\item{weights}{the name of a variable in the data frame that should be used
as a frequency weight.}

\item{offset}{offset the name of a variable, or vector of variable names, in the data frame that should be used 
as an offset (i.e., included but forced to have a coefficient of 1).}

\item{method}{a method to use for optimization in the maximum likelihood estimation. For options, see \code{\link[maxLik]{maxLik}},}

\item{max.iters}{the maximum number of iterations to allow the optimization method to perform,}

\item{start.vals}{an optional vector of starting values for the regression coefficients}

\item{print.level}{determines the level of verbosity for printing details of the optimization as it is computed. A value of 0 does not print out any information, a value of 1 prints minimal information, and a value of 2 prints the most information.}
}
\description{
Function for estimating a random parameter negative binomial with the ability to specify if the NB-1, NB-2, or NB-P should be used
}
\examples{
\donttest{

## Random Parameters Negative Binomial model (NB-1)
data("washington_roads")
nb1.rp <- rpnb(Total_crashes ~ - 1 + lnlength + lnaadt,
               rpar_formula = ~ speed50,
               data = washington_roads,
               ndraws = 100,
               correlated = FALSE,
               rpardists = c(intercept="u", speed50="t"),
               form = 'nb1',
               method = "bfgs",
               print.level = 2)

summary(nb1.rp)

## Random Parameters Negative Binomial model (NB-2)
nb2.rp <- rpnb(Total_crashes ~ - 1 + lnlength + lnaadt,
               rpar_formula = ~ speed50,
               data = washington_roads,
               ndraws = 100,
               correlated = TRUE,
               form = 'nb2',
               method = "bfgs",
               print.level = 1)

summary(nb2.rp)

## Random Parameters Negative Binomial model (NB-P)
nbp.rp <- rpnb(Total_crashes ~ - 1 + lnlength + lnaadt,
               rpar_formula = ~ speed50,
               data = washington_roads,
               ndraws = 100,
               correlated = TRUE,
               form = 'nbp',
               method = "bfgs",
               print.level = 1)

summary(nbp.rp)}
}
