% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_flexCountReg.R
\name{summary.flexCountReg}
\alias{summary.flexCountReg}
\title{Custom summary method for models with bootstrapped standard errors}
\usage{
\method{summary}{flexCountReg}(object, confint_level = 0.95, digits = 3)
}
\arguments{
\item{object}{A model object with bootstrapped standard errors.}

\item{confint_level}{A numeric value between 0 and 1 indicating the confidence level for confidence intervals. Default is 0.95.}

\item{digits}{Number of digits (decimal places) to round to. Default is 3.}
}
\description{
Custom summary method for models with bootstrapped standard errors
}
\details{
Bootstrapped standard errors are currently only implemented in \code{\link{pwiebreg}} and \code{\link{rppoisson}}.
}
\examples{
# Poisson-Weibull
pw_rp <- pwiebreg(Total_crashes ~ lnlength + lnaadt,
                                 data = washington_roads,
                                 ndraws = 10,
                                 bootstraps = 10)
summary(pw_rp)

}
